% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-tools_adjustStandAges.R
\name{adjustCohortAges}
\alias{adjustCohortAges}
\title{Adjust cohort ages}
\usage{
adjustCohortAges(
  cohortAges,
  yearInput,
  yearOutput,
  disturbanceEvents = NULL,
  delay = NULL,
  default = NULL,
  warn = TRUE
)
}
\arguments{
\item{cohortAges}{data.table. Table of cohort ages
with an ID column and the numeric columns 'age'.
Optionally include a 'delay' column with regeneration delays for each cohort.
A standard delay can also be provided with the \code{delay} argument.
Optionally include a 'default' column with the default age for the cohort
if it cannot be otherwise calculated.
A default age can also be provided with the \code{default} argument.}

\item{yearInput}{integer. The year that the 'ages' column in \code{cohortAges} represents.}

\item{yearOutput}{integer. The year that cohort ages must be adjusted to.}

\item{disturbanceEvents}{data.table. Optional.
Table of disturbance events with the ID column in \code{cohortAges}
and another column 'year' of when the disturbance occurred.
This can also include rows for years that cohorts were first established.}

\item{delay}{integer. Optional. Regeneration delay after a disturbance event.}

\item{default}{integer. A default age for cohorts is otherwise unknown.
If \code{yearOutput} precedes \code{yearInput},
the age cannot be calculated for cohorts that have disturbances between
\code{yearOutput} and \code{yearInput} but none before \code{yearOutput}.
\code{default} will be assigned in these cases.
NOTE: if a cohort has an age lesser than the difference in these years,
it is assumed that a disturbance event must have occurred when age == 0.
\code{default} will be assigned in these cases.}

\item{warn}{logical. Warn if ages cannot be calculated
or if the provided disturbances do not match the input data.}
}
\value{
\code{cohortAges} with ages adjusted to \code{yearOutput}.
}
\description{
Adjust cohort ages to represent a year before or after a given date.
Optionally include disturbance events that reset ages to year 0.
}
