% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-tools_adjustStandAges.R
\name{adjustStandAges}
\alias{adjustStandAges}
\title{Adjust stand ages}
\usage{
adjustStandAges(
  standAges,
  yearInput,
  yearOutput,
  disturbanceEvents = NULL,
  defaultAge = NULL,
  delay = NULL
)
}
\arguments{
\item{standAges}{data.table. Table of stand ages
with an ID column and the numeric columns 'age'.
Optionally include a 'delay' column with regeneration delays for each stand.
A standard delay can also be provided with the \code{delay} argument.}

\item{yearInput}{integer. The year that the 'ages' column in \code{standAges} represents.}

\item{yearOutput}{integer. The year that stand ages must be adjusted to.}

\item{disturbanceEvents}{data.table. Optional.
Table of disturbance events with the ID column in \code{standAges}
and another column 'year' of when the disturbance occurred.}

\item{defaultAge}{integer. A default age for stands before the first disturbance.
If \code{yearOutput} precedes \code{yearInput}, stands with ages
lesser than the difference in these years may be calculated to have an age <0,
suggesting that a disturbance must have had occurred when age == 0.
if \code{disturbanceEvents} includes an event before this date,
stand age will be calculated from this date.
Otherwise, negative ages are replaced with \code{defaultAge}.}

\item{delay}{integer. Optional. Regeneration delay after a disturbance event.}
}
\value{
\code{standAges} with ages adjusted to \code{yearOutput}.
}
\description{
Adjust stand ages to represent a year before or after a given date.
Optionally include disturbance events that reset ages to year 0.
}
