% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{barPlot}
\alias{barPlot}
\title{\code{barPlot}}
\usage{
barPlot(cbmPools, masterRaster)
}
\arguments{
\item{cbmPools}{TODO}

\item{masterRaster}{A \code{SpatRaster} object to use as the template}
}
\value{
TODO
}
\description{
\code{barPlot}
}
