% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{biomassTurnoverMatrix}
\alias{biomassTurnoverMatrix}
\title{Biomass turnover matrix}
\usage{
biomassTurnoverMatrix(turnoverParam, PoolCount)
}
\arguments{
\item{turnoverParam}{DESCRIPTION NEEDED}

\item{PoolCount}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Biomass turnover matrix
}
