% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_growthCurves.R
\name{biomProp}
\alias{biomProp}
\title{Proportions of total tree biomass in stemwood, bark, branches, and foliage}
\usage{
biomProp(
  x,
  type = "volume",
  bTable6 = NULL,
  bTable7 = NULL,
  bTable6tb = NULL,
  bTable7tb = NULL
)
}
\arguments{
\item{x}{\code{vector} gross merchantable volume per hectare (\eqn{m^3/ha}) or
total biomass (\eqn{tonnes/ha})}

\item{type}{\code{character} specifies if the \code{x} represents gross merchantable
volume per hectare ("volume") or total biomass ("biomass").}

\item{bTable6}{Boudewyn et al. (2007) proportion model parameters.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table6.csv}.}

\item{bTable7}{Boudewyn et al. (2007) caps on proportion models.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table7.csv}.}

\item{bTable6tb}{Boudewyn et al. (2007) an alternative set of proportion model parameters
to use when total biomass per hectare in tonnes (tb = stem wood + stem bark + branches + foliage)
is the independent variable instead of gross merchantable volume per hectare.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table6_tb.csv}.}

\item{bTable7tb}{Boudewyn et al. (2007) an alternative set of caps on proportion models
to use in conjunction with the alternative set of proportion models
when total biomass per hectare in tonnes (tb = stem wood + stem bark + branches + foliage)
is the independent variable instead of gross merchantable volume per hectare.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table7_tb.csv}.}
}
\value{
four-column matrix will columns corresponding to \eqn{p_{stemwood}}, \eqn{p_{bark}},
\eqn{p_{branches}}, and \eqn{p_{foliage}}
}
\description{
Implements equations 4-7 of Boudewyn et al. (2007), used to determine the proportions
of total tree biomass in stemwood, bark, branches, and foliage
(\eqn{p_{stemwood}}, \eqn{p_{bark}}, \eqn{p_{branches}}, \eqn{p_{foliage}}, respectively),
using parameters \eqn{a}, \eqn{b} from Table 6 and volume-proportion caps
from Table 7.
}
\details{
TODO: will eventually add species, ecozone
}
\references{
Boudewyn, P., Song, X., Magnussen, S., & Gillis, M. D. (2007). Model-based, volume-to-biomass
conversion for forested and vegetated land in Canada (BC-X-411). Natural Resource Canada,
Pacific Forestry Centre. \url{https://cfs.nrcan.gc.ca/pubwarehouse/pdfs/27434.pdf}
}
