% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{calcC}
\alias{calcC}
\title{Sum carbon for \code{totalCarbon} or \code{aboveGround} or \code{belowGround}}
\usage{
calcC(cbmPools, poolToSum, masterRaster)
}
\arguments{
\item{cbmPools}{DESCRIPTION NEEDED}

\item{poolToSum}{DESCRIPTION NEEDED}

\item{masterRaster}{A \code{SpatRaster} object to use as the template}
}
\value{
DESCRIPTION NEEDED
}
\description{
Sum carbon for \code{totalCarbon} or \code{aboveGround} or \code{belowGround}
}
