% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{calcTurnoverRates}
\alias{calcTurnoverRates}
\title{Calculate turnover rates}
\usage{
calcTurnoverRates(turnoverRates, spatialUnitIds, spatialUnits)
}
\arguments{
\item{turnoverRates}{DESCRIPTION NEEEDED}

\item{spatialUnitIds}{DESCRIPTION NEEEDED}

\item{spatialUnits}{DESCRIPTION NEEEDED}
}
\value{
extracts the turnover rates for the specific SPU you are in.
These are used to in the core module to calculate the specific rates, which
are the used to calculate Net Primary Productivity (NPP) both in the core
module and in the next function.
}
\description{
For all spatial units in the simulation, calculate turnover rates.
}
