% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonara-package.R
\docType{package}
\name{carbonara-package}
\alias{carbonara}
\alias{carbonara-package}
\title{carbanara}
\description{
Implementation of several components of the Carbon Budget Model of the
    Canadian Forest Service (v3).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://carbonara.predictiveecology.org/}
  \item \url{https://github.com/PredictiveEcology/carbonara}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID}) [contributor]

Authors:
\itemize{
  \item Celine Boisvenue \email{celine.boisvenue@canada.ca}
  \item Ian Eddy \email{ian.eddy@canada.ca}
}

}
