% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_mapTotalCarbon.R
\name{cbm4MapTotalCarbon}
\alias{cbm4MapTotalCarbon}
\title{CBM4: \code{mapTotalCarbon}}
\usage{
cbm4MapTotalCarbon(cbm4_results, years = NULL, yearStart = 1)
}
\arguments{
\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{\link[CBM4r]{cbm4_results_processor}}.}

\item{years}{integer. Year(s) of simulation results.}

\item{yearStart}{integer. Simulation start year.}
}
\value{
\code{ggplot}
}
\description{
Map total carbon across a study area.
}
