% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_plotPoolProportions.R
\name{cbm4PlotPoolProportions}
\alias{cbm4PlotPoolProportions}
\title{CBM4: \code{plotPoolProportions}}
\usage{
cbm4PlotPoolProportions(cbm4_results, years = NULL, yearStart = 1)
}
\arguments{
\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{\link[CBM4r]{cbm4_results_processor}}.}

\item{years}{integer. Year(s) of simulation results.}

\item{yearStart}{integer. Simulation start year.}
}
\value{
\code{ggplot}
}
\description{
CBM4: \code{plotPoolProportions}
}
