% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-CBMsources.R
\name{CBMsourceExtractToRast}
\alias{CBMsourceExtractToRast}
\title{CBM source: Extract to raster}
\usage{
CBMsourceExtractToRast(sourceID, templateRast, ...)
}
\arguments{
\item{sourceID}{Source identifier.}

\item{templateRast}{SpatRaster. Raster template.}

\item{...}{additional arguments to \code{link[reproducible]{prepInputs}}}
}
\description{
Prepare a data source from the \code{\link{CBMsources}} table
with \code{link[reproducible]{prepInputs}},
then align the data with a template raster and extract values for each cell
with \code{link{extractToRast}}.
}
