% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-CBMsources.R
\name{CBMsourcePrepInputs}
\alias{CBMsourcePrepInputs}
\title{CBM source: Prep inputs}
\usage{
CBMsourcePrepInputs(
  sourceID,
  inputPath = getOption("spades.inputPath", "."),
  ...
)
}
\arguments{
\item{sourceID}{Source identifier.}

\item{inputPath}{character.
Path of download destination directory.
Downloads will be sorted into subdirectories by data provider and region.}

\item{...}{additional arguments to \code{link[reproducible]{prepInputs}}}
}
\description{
Prepare a data source from the \code{\link{CBMsources}} table
with \code{link[reproducible]{prepInputs}}.
}
