% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-CBMsources.R
\docType{data}
\name{CBMsources}
\alias{CBMsources}
\title{CBM Data Sources}
\format{
A data table with 8 columns and 6 rows:
\describe{
\item{sourceID}{Source identifier.}
\item{provider}{Data provider or inventory identifier.}
\item{year}{Data year. The year that the data represents or was published.}
\item{region}{Spatial domain. Defaults to all of Canada.}
\item{attr}{Cohort or stand attribute defined by the source.}
\item{type}{Spatial data type: 'vector' or 'raster'.}
\item{source}{Data table of one or more downloadable data sources:
\itemize{
\item url: Download URL of target data source.
\item targetFile: File name of target data source.
\item layer: Vector or raster layer.
\item field: Vector attribute field.
\item subattr: If the source is comprised of multiple vector fields or raster layers,
this secondary attribute distinguishes each one.
}
}
...
}
}
\usage{
CBMsources
}
\description{
Data sources for CBM simulations. These can be read with \code{CBMsourcePrepInputs}.
}
\keyword{datasets}
