% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBMutils-package.R
\docType{package}
\name{CBMutils-package}
\alias{CBMutils}
\alias{CBMutils-package}
\title{\code{CBMutils}}
\description{
Utilities for modelling carbon in R and SpaDES based on the Carbon Budget Model
of the Canadian Forest Service v3 (CBM-CFS3).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PredictiveEcology/CBMutils}
}

}
\author{
\strong{Maintainer}: Céline Boisvenue \email{celine.boisvenue@nrcan-rncan.gc.ca}

Authors:
\itemize{
  \item Céline Boisvenue \email{celine.boisvenue@nrcan-rncan.gc.ca}
  \item Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})
  \item Eliot McIntire \email{eliot.mcintire@nrcan-rncan.gc.ca}
}

Other contributors:
\itemize{
  \item Ian Eddy \email{ian.eddy@nrcan-rncan.gc.ca} [contributor]
  \item Camille Giuliano \email{camsgiu@gmail.com} [contributor]
  \item Susan Murray \email{murray.e.susan@gmail.com} [contributor]
  \item Dominique Caron \email{dominique.caron@nrcan-rncan.gc.ca} [contributor]
}

}
\keyword{internal}
