% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProportionalMatrices.R
\name{checkProp}
\alias{checkProp}
\title{check that the proportions add up to 1}
\usage{
checkProp(DTofMatrices, tarCols)
}
\arguments{
\item{DTofMatrices}{DESCRIPTION NEEDED}

\item{tarCols}{DESCRIPTION NEEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
check that the proportions add up to 1
}
