% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postSimCtransactionCheckFunctions.R
\name{checkTransactions}
\alias{checkTransactions}
\title{Verify and check transactions}
\usage{
checkTransactions(sim, n = 3)
}
\arguments{
\item{sim}{A simulation (\code{simList}) object.}

\item{n}{number of pixelGroups (i.e., size of subset) to check}
}
\description{
Post-simulations, this function takes the last two years of simulations,
and verifies that the C transfers are as expected.
It randomly selects three \code{pixelGroups}, calculated the expected transactions from time
\code{end(sim)} and time \code{end(sim) - 1}, and compares them to the last year of simulations.
Cases of disturbed, non disturbed, sw, hw, young and old stands were tested.
}
\section{Structure}{

\enumerate{
\item Read-in data
\item randomly select 3 pixelGroups
\item read-in carbon pools at time(sim) and time(sim) - 1 for the three pixelGroup
\item pull-out matrices applied to these three pixelGroups between t0 and t1
\item matrices into a data table
\item prep pools at t0
\item merge three pools at t0 with their respective transaction matrices ids
\item apply all matrices (8.1:8.9)
}

TODO: I am passing a whole sim list here for simplification, but only four
objects are needed from the simList: $cbmPools, $pixelKeep, $opMatrixCBM,
$allProcesses, and $pooldef.

TODO: there are still a bunch of checks that need to be changed to assertions (if statements?)
}

