% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{computeBioTurnoverMatrices}
\alias{computeBioTurnoverMatrices}
\title{Compute Turnover Matrices}
\usage{
computeBioTurnoverMatrices(turnoverParameters, PoolCount)
}
\arguments{
\item{turnoverParameters}{DESCRIPTION NEEDED}

\item{PoolCount}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Compute Turnover Matrices
}
