% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{computeDomDecayMatrices}
\alias{computeDomDecayMatrices}
\title{Compute all dom decay matrices in coordinate matrix format}
\usage{
computeDomDecayMatrices(decayRates, decayParameters, PoolCount)
}
\arguments{
\item{decayRates}{matrix of decay rates column 1 is the key for the values
in columns \code{1:n} and columns \code{1:n} are the DOM pool specific decay rates.}

\item{decayParameters}{table of CBM decay parameters}

\item{PoolCount}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
The first column in the specified \code{decayRates} parameter acts as the key to each matrix.
}
