% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{computeDomTurnoverMatrices}
\alias{computeDomTurnoverMatrices}
\title{Compute DOM turnover matrices}
\usage{
computeDomTurnoverMatrices(turnoverParameters, PoolCount)
}
\arguments{
\item{turnoverParameters}{DESCRIPTION NEEDED}

\item{PoolCount}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Compute DOM turnover matrices
}
