% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeGrowthAndDeclineMatrices}
\alias{ComputeGrowthAndDeclineMatrices}
\title{Compute growth and decline matrices}
\usage{
ComputeGrowthAndDeclineMatrices(growthIncrements, decline)
}
\arguments{
\item{growthIncrements}{Matrix of growth increments}

\item{decline}{Matrix containing proportion decline for each pool}
}
\description{
DESCRIPTION NEEDED
}
