% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeGrowthAndDeclineMatrices2}
\alias{ComputeGrowthAndDeclineMatrices2}
\title{Compute growth and decline matrices}
\usage{
ComputeGrowthAndDeclineMatrices2(
  growthIncrements,
  ages,
  gcids,
  pools,
  rootParameters,
  turnoverParams,
  biomassToCarbonRate,
  swMult = 1,
  hwMult = 1
)
}
\arguments{
\item{growthIncrements}{a hash table of growth increments by \code{gcid}, by \code{age}}

\item{ages}{the stand age, (the inventory age) stands will be simulated to this age in the final pass}

\item{gcids}{integer vector of growth curve ids}

\item{pools}{Matrix (DESCRIPTION NEEDED)}

\item{rootParameters}{\code{data.frame} specifying root parameters}

\item{turnoverParams}{`data.frame specifying turnover parameters}

\item{biomassToCarbonRate}{numeric DESCRIPTION NEEDED}

\item{swMult}{softwood multiplier? (DESCRITION NEEDED)}

\item{hwMult}{hardwood multiplier? (DESCRITION NEEDED)}
}
\description{
DESCRIPTION NEEDED
}
