% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeGrowthCoordinateMatrices}
\alias{ComputeGrowthCoordinateMatrices}
\title{Compute growth coordinate matrices}
\arguments{
\item{pn}{\code{PoolNames} object}

\item{growthIncrements}{Matrix of growth increments}
}
\description{
Transforms growth matrix into coordinate format matrix in terms of CBM pools.
}
