% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeGrowthIncrements}
\alias{ComputeGrowthIncrements}
\title{Compute growth increments}
\usage{
ComputeGrowthIncrements(
  growthIncrements,
  ages,
  gcids,
  pools,
  rootParameters,
  biomassToCarbonRate,
  swMult = 1,
  hwMult = 1
)
}
\arguments{
\item{growthIncrements}{a hash table of growth increments by \code{gcid}, by \code{age}}

\item{ages}{the stand age, (the inventory age) stands will be simulated to this age in the final pass}

\item{gcids}{DESCRIPTION NEEDED}

\item{pools}{DESCRIPTION NEEDED}

\item{rootParameters}{\code{data.frame} (DESCRIPTION NEEDED)}

\item{biomassToCarbonRate}{DESCRIPTION NEEDED}

\item{swMult}{softwood multiplier? (DESCRIPTION NEEDED)}

\item{hwMult}{hardwood multiplier? (DESCRIPTION NEEDED)}
}
\description{
Compute the total growth (above ground and belowground) increment matrices for all
stands based on age and hashtable of growth increments by growth curve id.
}
