% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeOvermatureDecline}
\alias{ComputeOvermatureDecline}
\title{Compute overmature decline}
\usage{
ComputeOvermatureDecline(growthIncrements, turnoverParams)
}
\arguments{
\item{growthIncrements}{Matrix of growth increments}

\item{turnoverParams}{\code{data.frame} specifying turnover parameters}
}
\description{
Compute flows to dom pools that occur on a growth curve decline.
}
