% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeOverMatureDeclineCoordinateMatrices}
\alias{ComputeOverMatureDeclineCoordinateMatrices}
\title{Compute overmature decline coordinate matrices}
\arguments{
\item{pn}{\code{PoolNames} object}

\item{decline}{Matrix containing proportion decline for each pool}
}
\description{
Transforms overmature decline matrix into coordinate format matrix in terms of CBM pools.
}
