% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{computeSlowDecayMatrices}
\alias{computeSlowDecayMatrices}
\title{Compute slow decay matrices}
\usage{
computeSlowDecayMatrices(decayRates, decayParameters, PoolCount)
}
\arguments{
\item{decayRates}{DESCRIPTION NEEDED}

\item{decayParameters}{DESCRIPTION NEEDED}

\item{PoolCount}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Compute slow decay matrices
}
