% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{computeSlowMixingMatrix}
\alias{computeSlowMixingMatrix}
\title{Compute slow mixing matrix}
\usage{
computeSlowMixingMatrix(slowMixingRate, PoolCount)
}
\arguments{
\item{slowMixingRate}{DESCRIPTION NEEDED}

\item{PoolCount}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Compute slow mixing matrix
}
