% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_growthCurves.R
\name{convertM3biom}
\alias{convertM3biom}
\title{Calculate biomass from gross merchantable volume}
\usage{
convertM3biom(
  meta,
  gCvalues,
  spsMatch,
  ecozones,
  bTable3 = "https://nfi.nfis.org/resources/biomass_models/appendix2_table3.csv",
  bTable4 = "https://nfi.nfis.org/resources/biomass_models/appendix2_table4.csv",
  bTable5 = "https://nfi.nfis.org/resources/biomass_models/appendix2_table5.csv",
  bTable6 = "https://nfi.nfis.org/resources/biomass_models/appendix2_table6.csv",
  bTable7 = "https://nfi.nfis.org/resources/biomass_models/appendix2_table7.csv"
)
}
\arguments{
\item{meta}{Growth curve metadata}

\item{gCvalues}{Growth curve volume values for each age}

\item{spsMatch}{Species associated with growth curve}

\item{ecozones}{Ecozone associated with the growth curve}

\item{bTable3}{Boudewyn et al. (2007) stem wood biomass model parameters for merchantable-sized trees.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table3.csv}.}

\item{bTable4}{Boudewyn et al. (2007) stem wood biomass model parameters for nonmerchantable-sized trees.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table4.csv}.}

\item{bTable5}{Boudewyn et al. (2007) stem wood biomass model parameters for sapling-sized trees.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table5.csv}.}

\item{bTable6}{Boudewyn et al. (2007) proportion model parameters.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table6.csv}.}

\item{bTable7}{Boudewyn et al. (2007) caps on proportion models.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table7.csv}.}
}
\value{
three-column matrix with columns corresponding to biomass (\eqn{T/ha}) for
total merchantable, foliage, and other.
}
\description{
Implements the flowchart from figure 3 of Boudewyn et al. (2007) to determined the
total above ground biomass (\eqn{T/ha}) from gross merchantable volume (\eqn{m^3/ha}).
}
\references{
Boudewyn, P., Song, X., Magnussen, S., & Gillis, M. D. (2007). Model-based, volume-to-biomass
conversion for forested and vegetated land in Canada (BC-X-411). Natural Resource Canada,
Pacific Forestry Centre. \url{https://cfs.nrcan.gc.ca/pubwarehouse/pdfs/27434.pdf}
}
