% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{cTransfer}
\alias{cTransfer}
\alias{cTransferDist}
\title{Calculate C transfer for disturbances and annual processes post-disturbance}
\usage{
cTransfer(standIn, transProp)

cTransferDist(standIn, transProp)
}
\arguments{
\item{standIn}{DESCRIPTION NEEDED}

\item{transProp}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Mismatch in C transfers when disturbance happens in C++ processing so bypassing it
C transfer functions: one for the disturbance (so small decimals errors in
matrices are corrected), and one for the annual processes.
}
