% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots.R
\name{cTransfersAlluvial}
\alias{cTransfersAlluvial}
\title{\code{cTransfersAlluvial}}
\usage{
cTransfersAlluvial(cTransfers, distMatrixID = NA, distName = NA, spu)
}
\arguments{
\item{cTransfers}{TODO}

\item{distMatrixID}{disturbance_matrix_id of the disturbance to plot. The user needs to specify this and/or disturbance name.}

\item{distName}{disturbance name you wish to plot. The user needs to specify this and/or disturbance matrix id.}

\item{spatial_unit_id}{spatial unit of the disturbance to plot. This must be provided.}
}
\value{
alluvialDist Alluvial plot of a disturbance in a specific spatial unit.
}
\description{
Maps proportions of carbon transfers across pools during a disturbance.
}
