% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_cumPoolsCreate.R
\name{cumPoolsCreate}
\alias{cumPoolsCreate}
\title{Create \code{cumPools} data.table}
\usage{
cumPoolsCreate(
  fullSpecies,
  gcMeta,
  userGcM3,
  thisAdmin,
  bTable3 = "https://nfi.nfis.org/resources/biomass_models/appendix2_table3.csv",
  bTable4 = "https://nfi.nfis.org/resources/biomass_models/appendix2_table4.csv",
  bTable5 = "https://nfi.nfis.org/resources/biomass_models/appendix2_table5.csv",
  bTable6 = "https://nfi.nfis.org/resources/biomass_models/appendix2_table6.csv",
  bTable7 = "https://nfi.nfis.org/resources/biomass_models/appendix2_table7.csv",
  bRateBiomassToCarbon = 0.5
)
}
\arguments{
\item{fullSpecies}{Study area species' names}

\item{gcMeta}{Growth curve metadata table}

\item{userGcM3}{Table of growth curve volume by age}

\item{thisAdmin}{Ecozone and spatial unit information table for the study area}

\item{bTable3}{Boudewyn et al. (2007) stem wood biomass model parameters for merchantable-sized trees.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table3.csv}.}

\item{bTable4}{Boudewyn et al. (2007) stem wood biomass model parameters for nonmerchantable-sized trees.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table4.csv}.}

\item{bTable5}{Boudewyn et al. (2007) stem wood biomass model parameters for sapling-sized trees.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table5.csv}.}

\item{bTable6}{Boudewyn et al. (2007) proportion model parameters.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table6.csv}.}

\item{bTable7}{Boudewyn et al. (2007) caps on proportion models.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table7.csv}.}

\item{bRateBiomassToCarbon}{numeric. Conversion factor of biomass to carbon.}
}
\value{
\code{cumPools} data.table
}
\description{
Create \code{cumPools} data.table
}
