% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_cumPoolsSmooth.R
\name{cumPoolsSmooth}
\alias{cumPoolsSmooth}
\title{Smooth the \code{cumPools} \code{data.table}}
\usage{
cumPoolsSmooth(
  cumPoolsRaw,
  colsToUse = c("totMerch", "fol", "other"),
  colsToUseNew = paste0(colsToUse, "_New")
)
}
\arguments{
\item{cumPoolsRaw}{Data.table with a numeric column called \verb{age`` and columns called }colsToUse`.}

\item{colsToUse}{A character vector of columns to smooth.}

\item{colsToUseNew}{A character vector of column names for the new smoothed columns}
}
\value{
#' A \code{data.table} with original columns plus new columns named `colsToUseNew``.
}
\description{
This uses the Chapman Richards equation to smooth the curves that are in \code{colsToUse}.
}
\author{
Celine Boisvenue and Eliot McIntire
}
