% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep_disturbanceRasters.R
\name{dataPrep_disturbanceRasters}
\alias{dataPrep_disturbanceRasters}
\title{CBM data prep: disturbanceRasters}
\usage{
dataPrep_disturbanceRasters(
  disturbanceRasters,
  eventID = NULL,
  templateRast = NULL,
  year = NULL
)
}
\arguments{
\item{disturbanceRasters}{terra \code{\link[terra]{SpatRaster}} or list of SpatRaster.
Must be named with 4 digit years such that a single raster layer can be accessed
for each disturbance year (e.g.  \code{disturbanceRasters[["2025"]]}).}

\item{eventID}{integer. Disturbance event ID.
All non-NA cells will be included as events of this type.
If NULL, the raster cell values must be the event IDs.}

\item{templateRast}{terra \code{\link[terra]{SpatRaster}}.
Optional template raster to align rasters with.}

\item{year}{digit or character.
One or more years to summarize disturbance events for.
If NULL, all available years are summarized.}
}
\value{
\code{disturbanceEvents}
data.table with integer columns 'pixelIndex', 'year', 'eventID'.
}
\description{
Process \code{disturbanceRasters} into a table of \code{disturbanceEvents}
for the CBM_core module.
Optionally resample to align with a template raster.
}
