% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep_disturbanceRasters.R
\name{dataPrep_disturbanceRastersURL}
\alias{dataPrep_disturbanceRastersURL}
\title{CBM data prep: disturbanceRastersURL}
\usage{
dataPrep_disturbanceRastersURL(disturbanceRastersURL, bandYears = NULL, ...)
}
\arguments{
\item{disturbanceRastersURL}{character.
URL of either an archive of raster files or a single raster file.}

\item{bandYears}{4 digit numeric or character years.
If the URL is of a single raster file,
provide the disturbance years that each raster band represents.}

\item{...}{additional arguments to reproducible \code{\link[reproducible]{preProcess}}}
}
\value{
\code{disturbanceRasters}.
If URL is an archive: a list of terra \code{\link[terra]{SpatRaster}}
where each item is named by the disturbance year.
If URL is a single raster file: a terra \code{\link[terra]{SpatRaster}}
where each raster band layer is named by the disturbance year.
}
\description{
Process \code{disturbanceRastersURL} into \code{disturbanceRasters}.
}
