% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{decayRate}
\alias{decayRate}
\title{Calculate the decay rate based on mean annual temperature and other parameters}
\usage{
decayRate(meanAnnualTemp, baseDecayRate, q10, tref, max)
}
\arguments{
\item{meanAnnualTemp}{scalar temperature in degrees Celsius}

\item{baseDecayRate}{scalar base decay rate constant}

\item{q10}{the scalar q10 value}

\item{tref}{the reference temperature}

\item{max}{the maximum allowed decay rate}
}
\value{
the scalar rate of decay
}
\description{
This is the ecological theory for decomposing being used in CBM.
}
