% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-DB_disturbances-types.R
\name{distList}
\alias{distList}
\title{CBM-CFS3 Disturbance types}
\usage{
distList(dbPath, localeID = 1)
}
\arguments{
\item{dbPath}{Path to CBM-CFS3 SQLite database file.}

\item{localeID}{CBM-CFS3 locale_id}
}
\value{
\code{data.table} with 'disturbance_type_tr' columns
'disturbance_type_id', 'name', 'description'
}
\description{
List CBM-CFS3 disturbance types.
}
