% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-DB_disturbances-types.R
\name{distMatch}
\alias{distMatch}
\title{CBM-CFS3 Disturbances Match}
\usage{
distMatch(
  distNames,
  nearMatches = TRUE,
  identical = !ask,
  ask = interactive(),
  listDist = NULL,
  dbPath = NULL,
  localeID = 1,
  ...
)
}
\arguments{
\item{distNames}{character.
Disturbance names to match with CBM-CFS3 database disturbance type names.}

\item{nearMatches}{logical. Allow for near matches; e.g. "clearcut" can match "clear-cut".}

\item{identical}{logical. Require identical matches.}

\item{ask}{logical.
If TRUE, prompt the user to choose the correct matches.
If FALSE, the function will look for a single match to each input.}

\item{listDist}{data.table. Optional. Result of a call to \code{\link{distList}}.
Table of disturbance types with columns
'disturbance_type_id', 'name', 'description'.}

\item{dbPath}{Path to CBM-CFS3 SQLite database file. Required if \code{listDist} is NULL.}

\item{localeID}{CBM-CFS3 locale_id}

\item{...}{additional arguments to \code{\link{.matchSelect}}}
}
\value{
\code{data.table} with columns 'disturbance_type_id', 'name', 'description'
}
\description{
Match disturbance names with CBM-CFS3 disturbance types.
}
