% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{domDecayMatrix}
\alias{domDecayMatrix}
\title{Compute a single dom decay matrix based on the specified table of decay rates}
\usage{
domDecayMatrix(decayRates, decayParameters, PoolCount)
}
\arguments{
\item{decayRates}{vector of decay rates (each element represents a dom pool)}

\item{decayParameters}{table of cbm decay parameters}

\item{PoolCount}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Compute a single dom decay matrix based on the specified table of decay rates
}
