% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{domDecayMatrixItem}
\alias{domDecayMatrixItem}
\title{Calculate a portion of the DOM decay matrix}
\usage{
domDecayMatrixItem(mat, decayRates, propToAtmosphere, src, dst, emission)
}
\arguments{
\item{mat}{the datatable}

\item{decayRates}{vector of annual decay rates by dom pool}

\item{propToAtmosphere}{vector of the proportions of decay emitted to the atmosphere as CO2 by this process}

\item{src}{the integer code for the dom pool being decayed}

\item{dst}{the integer code for the dom pool receiving non-emitted decayed matter}

\item{emission}{the integer code for the CO2 pool}
}
\value{
A modified copy of the input \code{mat}
}
\description{
Calculate a portion of the DOM decay matrix
}
