% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{domTurnOverMatrix}
\alias{domTurnOverMatrix}
\title{DOM turnover matrix}
\usage{
domTurnOverMatrix(turnoverParam, PoolCount)
}
\arguments{
\item{turnoverParam}{DESCRIPTION NEEDED}

\item{PoolCount}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
DOM turnover matrix
}
