% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadesCBMextraFunctions.R
\docType{data}
\name{.poolNames}
\alias{.poolNames}
\title{Disturbance matrix pool names}
\format{
An object of class \code{data.frame} with 25 rows and 2 columns.
}
\usage{
.poolNames
}
\description{
TODO: confirm these
}
\keyword{datasets}
