% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc-extractToRast.R
\name{extractToRast}
\alias{extractToRast}
\title{Extract to raster}
\usage{
extractToRast(input, templateRast, field = 1)
}
\arguments{
\item{input}{terra SpatRaster, one or more raster files, or sf polygons.}

\item{templateRast}{SpatRaster. Raster template.}

\item{field}{numeric or character. Vector field to extract.}
}
\value{
vector with a value for each cell of \code{templateRast}.
Data type matches input data type.
}
\description{
Extract values from a spatial data source for each cell of a template raster.
Resampling is done using "mode" resampling to preserve categorical values.
}
