% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_gcidsCreate.R
\name{gcidsCreate}
\alias{gcidsCreate}
\title{Standardized way to create gcids}
\usage{
gcidsCreate(...)
}
\arguments{
\item{...}{The data.table with ONLY the columns on which to determine unique gcids}
}
\description{
This uses the underscore delimiter between column names.
}
