% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-tools_gcLocatorCreate.R
\name{gcLocatorCreate}
\alias{gcLocatorCreate}
\title{Create gcLocator raster layer}
\usage{
gcLocatorCreate(siteProductivity, spuRaster, leadSpecies, gcidLookup)
}
\arguments{
\item{siteProductivity}{Productivity class raster layer}

\item{spuRaster}{spatial unit ID raster layer}

\item{leadSpecies}{leading species raster layer}

\item{gcidLookup}{Table with metadata for each gcid}
}
\value{
\code{gcLocator} SpatRaster
}
\description{
Builds the gcLocator raster from site productivity, spatial unit, and leading species matched to
a growth curve lookup table
}
