% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{getDecayRates}
\alias{getDecayRates}
\title{Get decay rates}
\usage{
getDecayRates(meanAnnualTemp, decayParameters, domPools)
}
\arguments{
\item{meanAnnualTemp}{scalar temperature in deg Celcius}

\item{decayParameters}{table of decay parameters for calculating the temperature dependent decay rate}

\item{domPools}{DESCRIPTION NEEDED}
}
\value{
the vector of decay rates
}
\description{
Returns a vector of decay rates where the indices of the vector are the DOM pools of CBM.
}
