% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getIdentityCoordinateMatrix}
\alias{getIdentityCoordinateMatrix}
\title{getIdentityCoordinateMatrix}
\usage{
getIdentityCoordinateMatrix(size)
}
\arguments{
\item{size}{Numeric indicating the number of rows and columns in the matrix.}
}
\description{
DESCRIPTION NEEDED
}
