% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTable}
\alias{getTable}
\title{getTable}
\usage{
getTable(filename, dbPath, sqlDir)
}
\arguments{
\item{filename}{DESCRIPTION NEEDED}

\item{dbPath}{DESCRIPTION NEEDED}

\item{sqlDir}{DESCRIPTION NEEDED}
}
\description{
DESCRIPTION NEEDED
}
