% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc-ageStepBackward.R
\name{gstat_krige_replace}
\alias{gstat_krige_replace}
\title{gstat::krige replace cells}
\usage{
gstat_krige_replace(
  inRast,
  cells,
  ignore = NULL,
  formula = z ~ 1,
  ...,
  agg.fact = 2,
  agg.fun = "median",
  agg.na.rm = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{ignore}{numeric. Raster values to exclude from input data.}

\item{formula}{Interpolation formula. See \code{\link[gstat]{krige}}.}

\item{...}{additional arguments to \code{\link[gstat]{krige}}.}

\item{agg.fact}{Aggregation factor for the input raster.
If >1, the raster will be aggregated before cell values
are extracted as input data points for interpolation.
This smooths the input data and speeds up the interpolation process.
See \code{\link[terra]{aggregate}}.}

\item{agg.fun}{Aggregation function.}

\item{agg.na.rm}{Remove NA cells in aggregation function.}
}
\description{
Replace cells in a raster by interpolating their values
from the other cells values with gstat::krige.
}
\keyword{internal}
