% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc-ageStepBackward.R
\name{gstat_replace}
\alias{gstat_replace}
\title{gstat replace cells}
\usage{
gstat_replace(
  inRast,
  cells,
  ignore = NULL,
  idw = TRUE,
  idp = 2,
  nmax = Inf,
  maxdist = Inf,
  ...,
  agg.fact = 1,
  agg.fun = "median",
  agg.na.rm = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{ignore}{numeric. Raster values to exclude from input data.}

\item{idw}{logical. Use \code{\link[gstat]{idw}} or \code{\link[gstat]{krige}}.}

\item{idp}{numeric. IDW power.}

\item{nmax}{numeric. The maximum number of nearest observations to use.}

\item{maxdist}{numeric. The maximum distance of observations to use.}

\item{...}{additional arguments to \code{\link[gstat]{idw}} or \code{\link[gstat]{krige}}.}

\item{agg.fact}{Aggregation factor for the input raster.
If >1, the raster will be aggregated before cell values
are extracted as input data points for interpolation.
This smooths the input data and speeds up the interpolation process.
See \code{\link[terra]{aggregate}}.}

\item{agg.fun}{Aggregation function.}

\item{agg.na.rm}{Remove NA cells when aggregating.}
}
\description{
Replace cells in a raster by interpolating their values
from the other cells values with gstat::idw or gstat::krige.
}
\keyword{internal}
