% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-DB_disturbances.R
\name{histDist}
\alias{histDist}
\title{CBM-CFS3 Historical Disturbances}
\usage{
histDist(spuIDs, localeID = 1, ask = FALSE, ...)
}
\arguments{
\item{spuIDs}{Spatial unit ID(s)}

\item{localeID}{CBM-CFS3 locale_id}

\item{ask}{logical.
If TRUE, prompt the user to choose the correct disturbance matches.
If FALSE, the function will look for exact name matches.}

\item{...}{arguments to \code{\link{spuDistMatch}}}
}
\description{
Identifies the stand-replacing wildfire disturbance in each spatial unit.
}
\details{
In all spatial units in Canada, the historical disturbance is set to fire.
Historical disturbances in CBM-CFS3 are used for "filling-up" the soil-related carbon pools.
Boudewyn et al. (2007) translate the m3/ha curves into biomass per ha in each of four pools:
total biomass for stem wood, total biomass for bark, total biomass for branches and total
biomass for foliage.
Biomass in coarse and fine roots, in aboveground- and belowground- very-fast, -fast, -slow,
in medium-soil, and in snags still needs to be estimated.
A stand-replacing fire disturbance is used in a disturb-grow cycle, where stands are disturbed
and regrown with turnover, overmature, decay, functioning until the dead organic matter pools
biomass values stabilize (+/- 10\%) (TODO: check this).
}
