% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc-ageStepBackward.R
\name{idw_replace}
\alias{idw_replace}
\title{IDW replace}
\usage{
idw_replace(
  cxyz,
  cells,
  ignore = NULL,
  idp = 2,
  nmax = Inf,
  parallel.cores = NULL,
  parallel.chunkSize = 25000,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{cxyz}{Table with xy coordinates, z values, and point IDs (c).}

\item{ignore}{numeric. Raster values to exclude from input data.}

\item{idp}{numeric. IDW power.}

\item{nmax}{numeric. The maximum number of nearest observations to use.}

\item{parallel.cores}{Number of cores to use in parallel processing.
See \code{\link[parallel]{mclapply}}.}

\item{parallel.chunkSize}{Number of cells to process in each parallel processing chunk.}

\item{...}{additional arguments to \code{\link[parallel]{mclapply}} or \code{\link[parallel]{mclapply}}.}
}
\description{
Replace values in a table by interpolating their values
from the other points values using IDW interpolation.
}
\keyword{internal}
