% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbanceMatrices.R
\name{loadDisturbanceMatrixIds}
\alias{loadDisturbanceMatrixIds}
\title{Load disturbance matrix IDs}
\usage{
loadDisturbanceMatrixIds(disturbanceMatrixValues, dbPools)
}
\arguments{
\item{disturbanceMatrixValues}{DESCRIPTION NEEDED}

\item{dbPools}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Load disturbance matrix IDs
}
