% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{m3ToBiomIncOnlyPlots}
\alias{m3ToBiomIncOnlyPlots}
\title{\code{m3ToBiomIncOnlyPlots}}
\usage{
m3ToBiomIncOnlyPlots(inc)
}
\arguments{
\item{inc}{TODO}
}
\value{
TODO
}
\description{
\code{m3ToBiomIncOnlyPlots}
}
