% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_m3ToBiomPlots.R
\name{m3ToBiomPlots}
\alias{m3ToBiomPlots}
\title{Makes plots of all columns for each id_col ID}
\usage{
m3ToBiomPlots(inc, id_col = "gcids", title)
}
\arguments{
\item{inc}{increment data}

\item{id_col}{column by which to group plots}

\item{title}{Title of plots}
}
\description{
Makes plots of all columns for each id_col ID
}
