% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadesCBMextraFunctions.R
\name{m3ToVolCheckPlots}
\alias{m3ToVolCheckPlots}
\title{m3ToVolCheckPlots}
\usage{
m3ToVolCheckPlots(sim)
}
\arguments{
\item{sim}{A \code{SpaDES} CBM simulation object (\code{simList}).}
}
\description{
Check the growth increments by above ground pool. This is
necessary because the Boudewyn et al parameters used to translate the
m3/ha into biomass/ha do not always work. Returns a list of plots, each
plot show the merch, fol, and other increments for a specific growth curve.
}
