% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_mapCarbon.R
\name{mapCarbon}
\alias{mapCarbon}
\title{\code{mapCarbon}}
\usage{
mapCarbon(pools, masterRaster, year = NULL)
}
\arguments{
\item{pools}{data.table. Table of pools for each cohort or for each pixel.
Must include the 'pixelIndex' column.}

\item{masterRaster}{\code{\link[terra]{SpatRaster}} template defining the study area.}

\item{year}{numeric. Year that the map represents.
If provided, it will be included in the plot title.}
}
\value{
\code{ggplot}
}
\description{
Map total carbon across a study area.
}
