% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_mapNPP.R
\name{mapNPP}
\alias{mapNPP}
\title{\code{mapNPP}}
\usage{
mapNPP(rastNPP, year = NULL)
}
\arguments{
\item{rastNPP}{SpatRaster. Pixel values represent NPP (t/ha).}

\item{year}{numeric. Year that the map represents.
If provided, it will be included in the plot title.}
}
\value{
\code{ggplot}
}
\description{
Map Net Primary Productivity (NPP) across a study area.
}
