% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_mapTotalCarbon.R
\name{mapTotalCarbon}
\alias{mapTotalCarbon}
\title{\code{mapTotalCarbon}}
\usage{
mapTotalCarbon(rastTC, year = NULL)
}
\arguments{
\item{rastTC}{SpatRaster. Pixel values represent total carbon (t/ha).}

\item{year}{numeric. Year that the map represents.
If provided, it will be included in the plot title.}
}
\value{
\code{ggplot}
}
\description{
Map total carbon across a study area.
}
