% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProportionalMatrices.R
\name{matrixDT}
\alias{matrixDT}
\title{Make a \code{data.table} out of the carbon transfer matrices}
\usage{
matrixDT(matricesIn, indicesIn)
}
\arguments{
\item{matricesIn}{DESCRIPTION NEEDED}

\item{indicesIn}{DESCRIPTION NEEDED}
}
\value{
a \code{data.table} object summarizing the carbon transfers
}
\description{
Make a \code{data.table} out of the carbon transfer matrices
}
