% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadesCBMextraFunctions.R
\name{NPP}
\alias{NPP}
\title{Calculate post-simulation net primary productivity}
\usage{
NPP(pools, pixelGroupSpu, year1, year2, turnoverRates)
}
\arguments{
\item{pools}{DESCRIPTION NEEDED}

\item{pixelGroupSpu}{DESCRIPTION NEEDED}

\item{year1, year2}{Consecutive start and end years (e.g., 1990, 1991)}

\item{turnoverRates}{DESCRIPTION NEEDED}
}
\description{
Calculate post-simulation net primary productivity
}
\examples{
\dontrun{
  spadesCBMout <- simInitAndSpades() ## TODO: fill this in
  NPP(pools = spadesCBMout$cbmPools,
      pixelGroupSpu = spadesCBMout$pixelGroupC[, .(pixelGroup, spatial_unit_id)],
      year1 = 1990, year2 = 1991,
      turnoverRates = spadesCBMout$turnoverRates)
}
}
