% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots.R
\name{NPPplot}
\alias{NPPplot}
\title{\code{NPPplot}}
\usage{
NPPplot(NPP, year, masterRaster, cohortGroupKeep)
}
\arguments{
\item{NPP}{NPP per cohort for the simulation.}

\item{year}{simulation year to plot}

\item{masterRaster}{A \code{SpatRaster} object to use as the template}

\item{cohortGroupKeep}{cohort to pixel information}
}
\value{
TODO
}
\description{
Maps the average net primary productivity (NPP) across the simulation years.
}
