% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadesCBMextraFunctions.R
\name{plotCarbonRasters}
\alias{plotCarbonRasters}
\title{Plotting pools}
\usage{
plotCarbonRasters(pixelkeep, cbmPools, poolsToPlot, years, masterRaster)
}
\arguments{
\item{pixelkeep}{DESCRIPTION NEEDED}

\item{cbmPools}{DESCRIPTION NEEDED}

\item{poolsToPlot}{DESCRIPTION NEEDED}

\item{years}{DESCRIPTION NEEDED}

\item{masterRaster}{DESCRIPTION NEEDED}
}
\description{
Plotting pools
}
\examples{
\dontrun{
# include 'totalCarbon' in poolsToPlot to add plot of total carbon
plotCarbonRasters(
  cbmPools = spadesCBMout$cbmPools,
  poolsToPlot = c("totalCarbon", "BelowGroundSlowSoil"),
  masterRaster = spadesCBMout$masterRaster,
  pixelkeep = spadesCBMout$pixelKeep,
  years = c(1990, 2000, 2005)
)
}
}
