% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_plotPoolProportions.R
\name{plotPoolProportions}
\alias{plotPoolProportions}
\title{\code{plotPoolProportions}}
\usage{
plotPoolProportions(pools)
}
\arguments{
\item{pools}{data.table. Table of pools for each cohort, pixel, or year.
Must include the 'year' column.}
}
\value{
\code{ggplot}
}
\description{
\code{plotPoolProportions}
}
