% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadesCBMextraFunctions.R
\docType{data}
\name{.poolnames}
\alias{.poolnames}
\alias{.poolids}
\title{Disturbance matrix pool names and ids}
\format{
An object of class \code{character} of length 26.

An object of class \code{environment} of length 26.
}
\usage{
.poolnames

.poolids
}
\description{
TODO: confirm these
}
\details{
NOTE: A reminder that indexing in R starts at 1, whereas in C++ it starts at 0.
Several C++ data structures do not include the Input category, so the indices there are
defined using \code{.poolids - 1}.
}
\keyword{datasets}
