% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputEcozones.R
\name{prepInputsEcozones}
\alias{prepInputsEcozones}
\title{prepInputs for ecozones}
\usage{
prepInputsEcozones(url, dPath, masterRaster)
}
\arguments{
\item{url}{Optional character string indicating the URL to download from.
If not specified, then no download will be attempted. If not entry
exists in the \code{CHECKSUMS.txt} (in \code{destinationPath}), an entry
will be created or appended to. This \code{CHECKSUMS.txt} entry will be used
in subsequent calls to
\code{prepInputs} or \code{preProcess}, comparing the file on hand with the ad hoc
\code{CHECKSUMS.txt}. See table in \code{\link[reproducible]{preProcess}}.}

\item{masterRaster}{A raster with NAs reprsenting "off study area" and otherwise
any arbitrary value for "in study area". Equivalent to
\code{rasterToMatch} argument in \code{\link[reproducible]{prepInputs}}}
}
\description{
prepInputs for ecozones
}
