% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputs.R
\name{prepInputsVRI}
\alias{prepInputsVRI}
\title{\code{prepInputsVRI}}
\usage{
prepInputsVRI(url, dPath, rasterToMatch)
}
\arguments{
\item{url}{A URL to the data}

\item{dPath}{destination path}

\item{rasterToMatch}{A raster with NAs representing "off study area" and otherwise
any arbitrary value for "in study area".
Equivalent to \code{rasterToMatch} argument in \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}.}
}
\description{
Read in the BC VRI, with growth curve information (from \code{ws3}), and creates a raster stack of
the age and \code{gcID}.
}
