% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputs.R
\name{prepInputsVRIage}
\alias{prepInputsVRIage}
\title{Read in the BC VRI, with growth curve information (from \code{ws3}), and creates a raster stack of
the age}
\usage{
prepInputsVRIage(
  VRIurl,
  dPath,
  rasterToMatch,
  targetFile,
  field = "PROJ_AGE_1"
)
}
\arguments{
\item{VRIurl}{a URL to the data}

\item{dPath}{destination path}

\item{rasterToMatch}{A \code{RasterLayer} with \code{NA} representing "off study area" and otherwise
any arbitrary value for "in study area".
Equivalent to \code{rasterToMatch} argument in \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}.}

\item{targetFile}{A \code{gdb.zip} file that is the inventory file.}

\item{field}{The age column name in the inventory file.}
}
\description{
Read in the BC VRI, with growth curve information (from \code{ws3}), and creates a raster stack of
the age
}
