% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query}
\alias{query}
\title{query}
\usage{
query(dbPath, sql)
}
\arguments{
\item{dbPath}{Path to sqlite database file.}

\item{sql}{SQL statement with to execute in the database.}
}
\description{
DESCRIPTION NEEDED
}
