% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Python_ReticulateFindPython.R
\name{reticulate_install_python_windows}
\alias{reticulate_install_python_windows}
\title{Reticulate: Install Python on Windows}
\usage{
reticulate_install_python_windows(
  version = NULL,
  useGit = TRUE,
  prompt = interactive(),
  pyenvRoot = tools::R_user_dir("CBMutils")
)
}
\arguments{
\item{version}{character. Python version string.}

\item{useGit}{logical. Allow \code{reticulate} to clone pyenv-win if Git is available.}

\item{prompt}{logical. Prompt user to approve download of pyenv-win tool.}

\item{pyenvRoot}{character. Path of location for install of the pyenv-win tool.
Defaults to the R user data directory.}
}
\description{
Download and use the latest version of the
\href{https://pypi.org/project/pyenv-win/}{pyenv-win}
Python version management tool to install Python on Windows.
}
\details{
If Git is available, \code{reticulate} will clone the
\href{https://github.com/pyenv-win/pyenv-win}{pyenv-win Github repository}
and use the tool to install Python.
If Git is not available, this function will instead download the latest version
directly from the repository and make it available to \code{reticulate}.
}
\keyword{internal}
