% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Python_ReticulateFindPython.R
\name{reticulate_python_path}
\alias{reticulate_python_path}
\title{Python interpreter path}
\usage{
reticulate_python_path(
  version = NULL,
  pyenvRoot = tools::R_user_dir("CBMutils"),
  pyenvOnly = FALSE
)
}
\arguments{
\item{version}{character. Python version or a comma separated list of version constraints.
See ?reticulate::virtualenv_starter 'version' argument}

\item{pyenvRoot}{character. Path to directory containing pyenv-win tool}

\item{pyenvOnly}{logical. Exclude versions not within a pyenv install directory}
}
\value{
character or NULL. IF found, a path to Python interpreter
}
\description{
Get path to Python interpreter, including installs at a given pyenv-win location
}
