% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-DB_disturbances.R
\name{seeDist}
\alias{seeDist}
\title{See disturbances}
\usage{
seeDist(
  EXN = TRUE,
  matrixIDs = NULL,
  cbm_defaults_db = NULL,
  cbm_exn_dir = NULL
)
}
\arguments{
\item{EXN}{logical. Use CBM-EXN CBM-CFS3 equivalent model data.}

\item{matrixIDs}{character. Optional. Subset disturbances by disturbance_matrix_id}

\item{cbm_defaults_db}{Path to CBM defaults SQLite database.}

\item{cbm_exn_dir}{Path to CBM-EXN parameters directory.}
}
\value{
List of \code{data.frame} named by disturbance_matrix_id
}
\description{
Retrieve disturbance source pools, sink pools, and the proportions transferred.
}
