% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadesCBMdb_dbReadRaw.R
\name{simCBMdbReadRaw}
\alias{simCBMdbReadRaw}
\title{simList: Read SpaDES CBM database}
\usage{
simCBMdbReadRaw(simCBM, year, table, ...)
}
\arguments{
\item{simCBM}{SpaDES.core \code{\link[SpaDES.core]{simList}} for a CBM simulation including the \strong{CBM_core} module.}

\item{year}{numeric. Year of simulation results.
Defaults to the current simulation year.}

\item{table}{character. Name of database table to read.
One of 'key', 'parameters', 'state', 'flux', or 'pools'.}

\item{...}{not used}
}
\value{
data.table
}
\description{
Read a raw data table from a simulation's \strong{CBM_core} SpaDES CBM database.
}
