% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_mapNPP.R
\name{simMapNPP}
\alias{simMapNPP}
\title{\code{simMapNPP}}
\usage{
simMapNPP(simCBM, year = NULL, useCache = TRUE)
}
\arguments{
\item{simCBM}{SpaDES.core \code{\link[SpaDES.core]{simList}} for a CBM simulation including the \strong{CBM_core} module.}

\item{year}{numeric. Year of simulation results. Defaults to the current simulation year.}

\item{useCache}{logical. Cache database query result to the database's
\code{cache} directory with \code{\link[reproducible]{Cache}}.}
}
\value{
\code{ggplot}
}
\description{
Map Net Primary Productivity (NPP) across a study area.
}
