% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_plotEmissionsProducts.R
\name{simPlotEmissionsProducts}
\alias{simPlotEmissionsProducts}
\title{\code{simPlotEmissionsProducts}}
\usage{
simPlotEmissionsProducts(simCBM, years = NULL, useCache = TRUE)
}
\arguments{
\item{simCBM}{SpaDES.core \code{\link[SpaDES.core]{simList}} for a CBM simulation including the \strong{CBM_core} module.}

\item{years}{numeric. Years of simulation results to summarize.
Required if \code{by = 'year'}.
Defaults to the full simulation time span.}

\item{useCache}{logical. Cache database query result to the database's
\code{cache} directory with \code{\link[reproducible]{Cache}}.}
}
\value{
\code{ggplot}
}
\description{
Plot yearly emissions and products.
}
