% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_plotPoolProportions.R
\name{simPlotPoolProportions}
\alias{simPlotPoolProportions}
\title{\code{simPlotPoolProportions}}
\usage{
simPlotPoolProportions(simCBM, years = NULL, useCache = TRUE)
}
\arguments{
\item{simCBM}{SpaDES.core \code{\link[SpaDES.core]{simList}} for a CBM simulation including the \strong{CBM_core} module.}

\item{years}{numeric. Simulation years to include in plot. Defaults to all simulation years.}

\item{useCache}{logical. Cache database query result to the database's
\code{cache} directory with \code{\link[reproducible]{Cache}}.}
}
\value{
\code{ggplot}
}
\description{
\code{simPlotPoolProportions}
}
