% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{slowDecayMatrix}
\alias{slowDecayMatrix}
\title{Slow decay martrix}
\usage{
slowDecayMatrix(decayRates, decayParameters, PoolCount)
}
\arguments{
\item{decayRates}{DESCRIPTION NEEDED}

\item{decayParameters}{DESCRIPTION NEEDED}

\item{PoolCount}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Slow decay martrix
}
