% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_mapTotalCarbon.R
\name{spadesCBMdbMapTotalCarbon}
\alias{spadesCBMdbMapTotalCarbon}
\title{spadesCBMdb \code{simMapTotalCarbon}}
\usage{
spadesCBMdbMapTotalCarbon(spadesCBMdb, masterRaster, year, useCache = TRUE)
}
\arguments{
\item{spadesCBMdb}{\strong{CBM_core} module outputs database path.}

\item{masterRaster}{\code{\link[terra]{SpatRaster}} template defining the study area.}

\item{year}{numeric. Year of simulation results.}

\item{useCache}{logical. Cache database query result to the database's
\code{cache} directory with \code{\link[reproducible]{Cache}}.}
}
\value{
\code{ggplot}
}
\description{
Map total carbon across a study area.
}
