% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadesCBMdb_dbReadRaw.R
\name{spadesCBMdbReadRaw}
\alias{spadesCBMdbReadRaw}
\title{Read CBM SpaDES CBM database}
\usage{
spadesCBMdbReadRaw(spadesCBMdb, year, table, ...)
}
\arguments{
\item{spadesCBMdb}{\strong{CBM_core} module outputs database path.}

\item{year}{numeric. Year of simulation results.}

\item{table}{character. Name of database table to read.
One of 'key', 'parameters', 'state', 'flux', or 'pools'.}

\item{...}{not used}
}
\value{
data.table
}
\description{
Read a raw data table from a \strong{CBM_core} SpaDES CBM database.
}
