% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadesCBMdb_dbReadSummary.R
\name{spadesCBMdbReadSummary}
\alias{spadesCBMdbReadSummary}
\title{Read CBM SpaDES CBM database summary}
\usage{
spadesCBMdbReadSummary(
  spadesCBMdb,
  summary,
  by = "cohortID",
  units = "t/ha",
  year = NULL,
  years = NULL,
  useCache = TRUE
)
}
\arguments{
\item{spadesCBMdb}{\strong{CBM_core} module outputs database path.}

\item{summary}{character. Name of summary table to collect.}

\item{by}{character. Name of field to summarize results by.
Data can be summarized by 'cohortID', 'pixelIndex', or 'year'}

\item{units}{character. 't/ha', 't', or 'Mt'.}

\item{year}{numeric. Year of simulation results.
Required if \verb{by = 'cohortID' or 'pixelIndex'}.}

\item{years}{numeric. Years of simulation results to summarize.
Required if \code{by = 'year'}.}

\item{useCache}{logical. Cache database query result to the database's
\code{cache} directory with \code{\link[reproducible]{Cache}}.}
}
\value{
data.table with key matching the \code{by} argument.
}
\description{
Read a data summary from a \strong{CBM_core} SpaDES CBM database.
}
