% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_plotPoolProportions.R
\name{spadesCBMPlotPoolProportions}
\alias{spadesCBMPlotPoolProportions}
\title{spadesCBMdb \code{simPlotPoolProportions}}
\usage{
spadesCBMPlotPoolProportions(spadesCBMdb, years, useCache = TRUE)
}
\arguments{
\item{spadesCBMdb}{\strong{CBM_core} module outputs database path.}

\item{years}{numeric. Simulation years to include in plot.}

\item{useCache}{logical. Cache database query result to the database's
\code{cache} directory with \code{\link[reproducible]{Cache}}.}
}
\value{
\code{ggplot}
}
\description{
spadesCBMdb \code{simPlotPoolProportions}
}
