% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{spatialRaster}
\alias{spatialRaster}
\title{Plot}
\usage{
spatialRaster(pixelkeep, cbmPools, poolsToPlot, years, masterRaster)
}
\arguments{
\item{pixelkeep}{DESCRIPTION NEEDED}

\item{cbmPools}{DESCRIPTION NEEDED}

\item{poolsToPlot}{DESCRIPTION NEEDED}

\item{years}{DESCRIPTION NEEDED}

\item{masterRaster}{A \code{SpatRaster} object to use as the template}
}
\value{
\code{RasterLayer} but invoked for its side-effect of plotting the rasters.
}
\description{
the \code{spatialPlot} function plots rasters directly, but does not return rasters, whereas
it does so here.
}
\seealso{
spatialPlot
}
