% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{spatialUnitDecayRates}
\alias{spatialUnitDecayRates}
\title{Spatial unit decay rates}
\usage{
spatialUnitDecayRates(climate, decayparameters, domPools)
}
\arguments{
\item{climate}{DESCRIPTION NEEDED}

\item{decayparameters}{DESCRIPTION NEEDED}

\item{domPools}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Spatial unit decay rates
}
