% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Spinup}
\alias{Spinup}
\title{Spinup}
\usage{
Spinup(
  pools,
  opMatrix,
  constantProcesses,
  growthIncrements,
  ages,
  gcids,
  historicdmids,
  lastPassdmids,
  delays,
  minRotations,
  maxRotations,
  returnIntervals,
  rootParameters,
  turnoverParams,
  biomassToCarbonRate,
  debug = FALSE
)
}
\arguments{
\item{pools}{Matrix (DESCRIPTION NEEDED)}

\item{opMatrix}{Matrix (DESCRIPTION NEEDED)}

\item{constantProcesses}{a list of constant process C dynamics matrices}

\item{growthIncrements}{a hash table of growth increments by \code{gcid}, by \code{age}}

\item{ages}{the stand age, (the inventory age) stands will be simulated to this age in the final pass}

\item{gcids}{the growth curve ids (referenced by \code{growthIncrements})}

\item{historicdmids}{(DESCRIPTION NEEDED)}

\item{lastPassdmids}{(DESCRIPTION NEEDED)}

\item{delays}{(DESCRIPTION NEEDED)}

\item{minRotations}{(DESCRIPTION NEEDED)}

\item{maxRotations}{(DESCRIPTION NEEDED)}

\item{returnIntervals}{(DESCRIPTION NEEDED)}

\item{rootParameters}{\code{data.frame} specifying root parameters}

\item{turnoverParams}{\code{data.frame} specifying turnover parameters}

\item{biomassToCarbonRate}{(DESCRIPTION NEEDED)}

\item{debug}{logical indicating whether to run spinup in debug mode.}
}
\description{
Spinup a landscape by running rotations of stand replacing disturbances repeatedly until the
pre-disturbance slow pools and the last rotation pre-disturbance slow pools are within a tolerance.
Stand then grows to inventory age, and is ready for CBM simulation.
}
