% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-DB_disturbances.R
\name{spuDistList}
\alias{spuDistList}
\title{CBM-CFS3 Spatial Unit Disturbances}
\usage{
spuDistList(
  EXN = TRUE,
  spuIDs = NULL,
  cbm_defaults_db = NULL,
  cbm_exn_dir = NULL,
  localeID = 1
)
}
\arguments{
\item{EXN}{logical. Use CBM-EXN CBM-CFS3 equivalent model data.}

\item{spuIDs}{Optional. Subset by spatial unit ID(s)}

\item{cbm_defaults_db}{Path to CBM defaults SQLite database.}

\item{cbm_exn_dir}{Path to CBM-EXN parameters directory.}

\item{localeID}{CBM-CFS3 locale_id}
}
\value{
\code{data.table} with 'disturbance_type_tr' columns
'disturbance_type_id', 'name', 'description'
and 'disturbance_matrix_association' columns
'spatial_unit_id' and 'disturbance_matrix_id'
}
\description{
List the disturbances possible in spatial units.
}
