% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StepPools}
\alias{StepPools}
\title{\code{StepPools}}
\usage{
StepPools(pools, opMatrix, flowMatrices)
}
\arguments{
\item{pools}{DESCRIPTION NEEDED}

\item{opMatrix}{DESCRIPTION NEEDED}

\item{flowMatrices}{DESCRIPTION NEEDED}
}
\description{
DESCRIPTION NEEDED
}
