% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StepPoolsRef}
\alias{StepPoolsRef}
\title{Step through pools, computing carbon transfers}
\usage{
StepPoolsRef(pools, opMatrix, flowMatrices)
}
\arguments{
\item{pools}{A numeric matrix describing the amount of carbon in each pool (column) across
multiple stands/pixel groups (rows).}

\item{opMatrix}{A numeric matrix with columns corresponding to flows, and rows corresponding
to stands/pixel groups. Values are indices into the \code{flowMatrices}.}

\item{flowMatrices}{A list of lists (or environments) containing all carbon flow matrices.}
}
\description{
Computes by row (stand/\code{pixelGroup}) across all columns (pools) in place.
}
