% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestModules.R
\name{suggestModules}
\alias{suggestModules}
\title{Suggest modules}
\usage{
suggestModules(sim, suggestedModules)
}
\arguments{
\item{sim}{DESCRIPTION NEEDED}

\item{suggestedModules}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Suggest modules
}
