% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc-extractToRast.R
\name{writeRasterWithValues}
\alias{writeRasterWithValues}
\title{Write raster with values}
\usage{
writeRasterWithValues(templateRast, values, filename, ...)
}
\arguments{
\item{templateRast}{terra \code{SpatRaster} template defining raster geometry.}

\item{values}{numeric or character. New raster values.}

\item{filename}{character. Output filename.}

\item{...}{arguments to \code{\link[terra]{writeRaster}}}
}
\description{
Replace the values in a \code{SpatRaster} before writing it to file.
}
