% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCharvestAndVRI.R
\name{ws3Build}
\alias{ws3Build}
\title{\code{ws3Build}}
\usage{
ws3Build(masterRaster, tsaDirs, years, pathsTifs)
}
\arguments{
\item{masterRaster}{A \code{SpatRaster} object to use as the template}

\item{tsaDirs}{character vector of paths to the TSA directories}

\item{years}{numeric or integer vector specifying the years for the data}

\item{pathsTifs}{character vector of paths for the geotiff files}
}
\description{
\code{ws3}, the harvest optimizer created by Greg Paradis provided the BC disturbance rasters
(harvest or harvest and fires) per TSA and not for the whole study area.
This function reads-in the file structure, pulls in all the TSA-level rasters,
knits them together and post-processes them to match the \code{rasterToMatch}.
}
