% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc-ageStepBackward.R
\name{ageStepBackward}
\alias{ageStepBackward}
\title{Age Step Backwards}
\usage{
ageStepBackward(
  ageRast,
  yearIn,
  yearOut,
  fill = TRUE,
  distEvents = NULL,
  idp = 2,
  nmax = 250,
  ...
)
}
\arguments{
\item{ageRast}{SpatRaster. Raster with numeric values of cohort ages.}

\item{yearIn}{numeric. Year that ages in \code{ageRast} represent.}

\item{yearOut}{numeric. Year that ages will be stepped backwards to.}

\item{fill}{logical. Fill disturbed areas and cells with ages <= 0 with values from surrounding cells.}

\item{distEvents}{data.table. Optional.
Table of disturbance events with columns "pixelIndex" and "year".
If provided, disturbances will be reversed and the disturbed areas
will be filled with \code{gstat_replace}.}

\item{idp}{numeric. IDW power.}

\item{nmax}{numeric. The maximum number of nearest observations to use.}

\item{...}{arguments to \code{\link{gstat_replace}}.}
}
\value{
\code{SpatRaster}
}
\description{
Step an age raster backwards in time.
Ages that are set as <=0 are replaced with surrounding ages >0 with \code{gstat_replace}.
}
