% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc-ageStepBackward.R
\name{ageStepBackward}
\alias{ageStepBackward}
\title{Age Step Backwards}
\usage{
ageStepBackward(
  ageRast,
  yearIn,
  yearOut,
  fill = TRUE,
  distEvents = NULL,
  formula = z ~ 1,
  agg.fact = 2,
  agg.fun = "median",
  ...
)
}
\arguments{
\item{ageRast}{SpatRaster. Raster with numeric values of cohort ages.}

\item{yearIn}{numeric. Year that ages in \code{ageRast} represent.}

\item{yearOut}{numeric. Year that ages will be stepped backwards to.}

\item{fill}{logical. Fill disturbed areas and cells with ages <= 0 with values from surrounding cells.}

\item{distEvents}{data.table. Optional.
Table of disturbance events with columns "pixelIndex" and "year".
If provided, disturbances will be reversed and the disturbed areas
will be filled with \code{gstat_krige_replace}.}

\item{formula}{Interpolation formula. See \code{\link[gstat]{krige}}.}

\item{agg.fact}{Aggregation factor for the input raster.
If >1, the raster will be aggregated before cell values
are extracted as input data points for interpolation.
This smooths the input data and speeds up the interpolation process.
See \code{\link[terra]{aggregate}}.}

\item{agg.fun}{Aggregation function.}

\item{...}{additional arguments to \code{\link[gstat]{krige}}.}
}
\value{
\code{SpatRaster}
}
\description{
Step an age raster backwards in time.
Ages that are set as <=0 are replaced with surrounding ages >0 with \code{gstat_krige_replace}.
}
