% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_cumPoolsCreateAGB.R
\name{alternativeParams}
\alias{alternativeParams}
\title{Extract alternative set of parameters to apply to convert total biomass into pool biomass.}
\usage{
alternativeParams(params, table, cols)
}
\arguments{
\item{params}{\code{data.table} containing a row for each curve with all required
parameters.}

\item{table}{\code{data.table} of the parameters.}

\item{cols}{\code{vector} of parameter column names.}
}
\value{
A single \code{data.table} containing a row for each curve with all required
parameters from \code{bTable6tb}, \code{bTable7tb}, and \code{tableMerch}.
}
\description{
When there is no match for the given ecozone, the parameters for a different
ecozone in the same province/territory is returned. If there is no match for
a given province/territory, the parameters for a different province/territory
in the same ecozone is returned. If there is no match for the given ecozone
and province/territory, the parameters for a different location is returned.
}
