% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthCurves.R
\name{b_m}
\alias{b_m}
\title{b_m}
\usage{
b_m(table3, vol)
}
\arguments{
\item{table3}{DESCRIPTION NEEDED}

\item{vol}{DESCRIPTION NEEDED}
}
\description{
eq1 gives the total stem wood biomass in metric tonnes/ha, when you give it
the gross merchantable volume/ha. Parameters a and b are in table3.
}
