% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_growthCurves.R
\name{biomProp}
\alias{biomProp}
\title{Proportions of total tree biomass in stemwood, bark, branches, and foliage}
\usage{
biomProp(table6, table7, x, type = "volume")
}
\arguments{
\item{table6}{\code{data.frame} corresponding to Table 6 from Boudewyn et al. (2007),
available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table6.csv}.
The alternative table 6 for equations using total biomass as independent variable
is available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table6_tb.csv}.}

\item{table7}{\code{data.frame} corresponding to Table 7 from Boudewyn et al. (2007),
available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table7.csv}.
The alternative table 7 for equations using total biomass as independent variable
is available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table7_tb.csv}.}

\item{x}{\code{vector} gross merchantable volume per hectare (\eqn{m^3/ha}) or
total biomass (\eqn{tonnes/ha})}

\item{type}{\code{character} specifies if the \code{x} represents gross merchantable
volume per hectare ("volume") or total biomass ("biomass").}
}
\value{
four-column matrix will columns corresponding to \eqn{p_{stemwood}}, \eqn{p_{bark}},
\eqn{p_{branches}}, and \eqn{p_{foliage}}
}
\description{
Implements equations 4-7 of Boudewyn et al. (2007), used to determine the proportions
of total tree biomass in stemwood, bark, branches, and foliage
(\eqn{p_{stemwood}}, \eqn{p_{bark}}, \eqn{p_{branches}}, \eqn{p_{foliage}}, respectively),
using parameters \eqn{a}, \eqn{b} from Table 6 (\code{table6}) and volume-proportion caps
from Table 7 (\code{table7}).
}
\details{
TODO: will eventually add species, ecozone
}
\references{
Boudewyn, P., Song, X., Magnussen, S., & Gillis, M. D. (2007). Model-based, volume-to-biomass
conversion for forested and vegetated land in Canada (BC-X-411). Natural Resource Canada,
Pacific Forestry Centre. \url{https://cfs.nrcan.gc.ca/pubwarehouse/pdfs/27434.pdf}
}
