% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthCurves.R
\name{biomProp}
\alias{biomProp}
\title{calculate the 4 proportions that should be returned}
\usage{
biomProp(table6, table7, vol)
}
\arguments{
\item{table6}{DESCRIPTION NEEDED}

\item{table7}{DESCRIPTION NEEDED}

\item{vol}{DESCRIPTION NEEDED}
}
\description{
will eventually add species, ecozone
vol = gross merchantable volume per ha
lvol = natural logarithm of (vol+5)
}
