% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_subsetTables.R
\name{boudewynSubsetTables}
\alias{boudewynSubsetTables}
\title{Subset Boudewyn tables to fit study area}
\usage{
boudewynSubsetTables(table, thisAdmin)
}
\arguments{
\item{table}{Boudewyn table to subset}

\item{thisAdmin}{Table defining study area with columns
\code{juris_id}: jurisdiction abbreviation(s); and
\code{ecozone}: ecozone ID(s).}
}
\value{
\code{smallTable} data.table
}
\description{
Subset Boudewyn tables to fit study area
}
