% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_subsetTables.R
\name{boudewynSubsetTables}
\alias{boudewynSubsetTables}
\title{Subset Boudewyn tables to fit study area}
\usage{
boudewynSubsetTables(table, thisAdmin, eco)
}
\arguments{
\item{table}{Boudewyn table to subset}

\item{thisAdmin}{study area attributes}

\item{eco}{ecozones in study area}
}
\value{
\code{smallTable} data.table
}
\description{
Subset Boudewyn tables to fit study area
}
