% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-tools_calcRootC.R
\name{calcRootC}
\alias{calcRootC}
\title{\code{calcRootC}}
\usage{
calcRootC(
  aboveGroundC,
  sw_hw,
  a_sw = 0.222,
  b_sw = 1,
  a_hw = 1.576,
  b_hw = 0.615,
  a_frp = 0.072,
  b_frp = 0.354
)
}
\arguments{
\item{aboveGroundC}{data.table with mass of carbon (tonnes/ha) in the Merch, Foliage and Other pools}

\item{sw_hw}{a boolean vector indicating if the cohort is softwood (0) or hardwood (1)}

\item{a_sw}{"a" value for softwood root biomass}

\item{b_sw}{"b" value for softwood root biomass}

\item{a_hw}{"a" value for hardwood root biomass}

\item{b_hw}{"b" value for hardwood root biomass}

\item{a_frp}{"a" value for fine root proportion}

\item{b_frp}{"b" value for fine root proportion}
}
\value{
data.table with mass of carbon (tonnes/ha) in coarseRoots and fineRoots pools.
}
\description{
\code{calcRootC} calculates the mass of carbon in roots pools from above ground pools
}
\references{
Li, Z., Kurz, W. A., Apps, M. J., & Beukema, S. J. (2003). Belowground biomass
dynamics in the Carbon Budget Model of the Canadian Forest Sector: recent improvements
and implications for the estimation of NPP and NEP. Canadian journal of forest
research, 33(1), 126-136.
}
