% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMatrices.R
\name{calcTurnoverRates}
\alias{calcTurnoverRates}
\title{Calculate turnover rates}
\usage{
calcTurnoverRates(turnoverRates, spatialUnitIds, spatialUnits)
}
\arguments{
\item{turnoverRates}{DESCRIPTION NEEEDED}

\item{spatialUnitIds}{DESCRIPTION NEEEDED}

\item{spatialUnits}{DESCRIPTION NEEEDED}
}
\value{
DESCRIPTION NEEEDED
}
\description{
For all spatial units in the simulation, calculate turnover rates.
}
