% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{carbonOutPlot}
\alias{carbonOutPlot}
\title{\code{carbonOutPlot}}
\usage{
carbonOutPlot(cbmPools, emissionsProducts, masterRaster)
}
\arguments{
\item{cbmPools}{TODO}

\item{emissionsProducts}{TODO}

\item{masterRaster}{TODO}
}
\value{
TODO
}
\description{
\code{carbonOutPlot}
}
