% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{carbonOutPlot}
\alias{carbonOutPlot}
\title{\code{carbonOutPlot}}
\usage{
carbonOutPlot(emissionsProducts)
}
\arguments{
\item{emissionsProducts}{TODO}
}
\value{
invoked for side effect of creating plot
}
\description{
\code{carbonOutPlot}
}
