% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{carbonOutPlot}
\alias{carbonOutPlot}
\title{carbonOutPlot}
\usage{
carbonOutPlot(cbmPools, emissionsProducts, masterRaster)
}
\arguments{
\item{cbmPools}{TODO}

\item{masterRaster}{TODO}
}
\value{
TODO
}
\description{
carbonOutPlot
}
