% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_mapNPP.R
\name{cbm4MapNPP}
\alias{cbm4MapNPP}
\title{\code{cbm4MapNPP}}
\usage{
cbm4MapNPP(cbm4_results, year, year1 = 1)
}
\arguments{
\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{CBM4::cbm4_results_processor}.}

\item{year}{numeric. Year of simulation results.}

\item{year1}{integer. Simulation start year.}
}
\value{
\code{ggplot}
}
\description{
Map net primary productivity (NPP) across a study area.
}
