% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_mapTotalCarbon.R
\name{cbm4MapTotalCarbon}
\alias{cbm4MapTotalCarbon}
\title{\code{cbm4MapTotalCarbon}}
\usage{
cbm4MapTotalCarbon(cbm4_results, year, year1 = 1)
}
\arguments{
\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{CBM4::cbm4_results_processor}.}

\item{year}{numeric. Year of simulation results.}

\item{year1}{integer. Simulation start year.}
}
\value{
\code{ggplot}
}
\description{
Map total carbon across a study area.
}
