% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_plotEmissionsProducts.R
\name{cbm4PlotEmissionsProducts}
\alias{cbm4PlotEmissionsProducts}
\title{\code{cbm4PlotEmissionsProducts}}
\usage{
cbm4PlotEmissionsProducts(cbm4_results, years = NULL, year1 = 1)
}
\arguments{
\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{CBM4::cbm4_results_processor}.}

\item{years}{numeric. Simulation years to include in plot. Defaults to all simulation years.}

\item{year1}{integer. Simulation start year.}
}
\value{
\code{ggplot}
}
\description{
Plot yearly emissions and products.
}
