% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonara-package.R
\docType{package}
\name{CBMutils-package}
\alias{CBMutils}
\alias{CBMutils-package}
\title{\code{CBMutils}}
\description{
Utilities for modelling carbon in R/SpaDES based on the Carbon Budget Model of the Canadian
Forest Service v3 (CBM-CFS3).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CBMutils.predictiveecology.org/}
  \item \url{https://github.com/PredictiveEcology/CBMutils}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

Authors:
\itemize{
  \item Celine Boisvenue \email{celine.boisvenue@canada.ca}
  \item Eliot McIntire \email{eliot.mcintire@canada.ca}
}

Other contributors:
\itemize{
  \item Ian Eddy \email{ian.eddy@canada.ca} [contributor]
}

}
