% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumPoolsCreateAGB.R
\name{convertAGB2pools}
\alias{convertAGB2pools}
\title{Convert total above ground biomass into 3 pools (\eqn{T/ha})}
\usage{
convertAGB2pools(oneCurve, table6, table7)
}
\arguments{
\item{oneCurve}{\code{data.frame} with at least four following columns: \code{canfi_species},
\code{ecozone}, \code{juris_id}, and \code{B}.}

\item{table6}{\code{data.frame} corresponding to Table 3 from Boudewyn et al. (2007),
available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table6_tb.csv}.}

\item{table7}{\code{data.frame} corresponding to Table 4 from Boudewyn et al. (2007),
available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table7_tb.csv}.}
}
\value{
three-column matrix with columns corresponding to biomass (\eqn{T/ha}) for
total merchantable, foliage, and other wood.
}
\description{
Implements the flowchart from figure 3 of Boudewyn et al. (2007) using an alternative
set of parameter to divide total above ground biomass (\eqn{T/ha}) into total merchantable
stemwood biomass (\eqn{T/ha}), foliage biomass (\eqn{T/ha}), and other wood biomass (\eqn{T/ha}).
}
\references{
Boudewyn, P., Song, X., Magnussen, S., & Gillis, M. D. (2007). Model-based, volume-to-biomass
conversion for forested and vegetated land in Canada (BC-X-411). Natural Resource Canada,
Pacific Forestry Centre. \url{https://cfs.nrcan.gc.ca/pubwarehouse/pdfs/27434.pdf}
}
