% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_cumPoolsCreateAGB.R
\name{convertAGB2pools}
\alias{convertAGB2pools}
\title{Convert total above ground biomass into 3 pools (\eqn{T/ha}).}
\usage{
convertAGB2pools(AGB, allParams)
}
\arguments{
\item{AGB}{\code{data.frame} with at least four following columns: \code{canfi_species},
\code{ecozone}, \code{juris_id}, and \code{B}.}

\item{allParams}{\code{data.frame} containing a row for each curve with all required
parameters of both \code{table6} and \code{table7} from Boudewyn et al. (2007) and parameters
on merchantability of stemwood.}
}
\value{
three-column matrix with columns corresponding to biomass (\eqn{T/ha}) for
total merchantable, foliage, and other wood.
}
\description{
Implements the flowchart from figure 3 of Boudewyn et al. (2007) using an alternative
set of parameter to divide total above ground biomass (\eqn{T/ha}) into total merchantable
stemwood biomass (\eqn{T/ha}), foliage biomass (\eqn{T/ha}), and other wood biomass (\eqn{T/ha}).
}
\references{
Boudewyn, P., Song, X., Magnussen, S., & Gillis, M. D. (2007). Model-based, volume-to-biomass
conversion for forested and vegetated land in Canada (BC-X-411). Natural Resource Canada,
Pacific Forestry Centre. \url{https://cfs.nrcan.gc.ca/pubwarehouse/pdfs/27434.pdf}
}
