% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthCurves.R
\name{convertM3biom}
\alias{convertM3biom}
\title{convertM3biom}
\usage{
convertM3biom(
  meta,
  gCvalues,
  spsMatch,
  ecozones,
  params3,
  params4,
  params5,
  params6
)
}
\arguments{
\item{meta}{DESCRIPTION NEEDED}

\item{gCvalues}{DESCRIPTION NEEDED}

\item{spsMatch}{DESCRIPTION NEEDED}

\item{ecozones}{DESCRIPTION NEEDED}

\item{params3}{DESCRIPTION NEEDED}

\item{params4}{DESCRIPTION NEEDED}

\item{params5}{DESCRIPTION NEEDED}

\item{params6}{DESCRIPTION NEEDED}
}
\description{
DESCRIPTION NEEDED
}
