% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthCurves.R
\name{convertM3biom}
\alias{convertM3biom}
\title{Calculate biomass from gross merchantable volume}
\usage{
convertM3biom(
  meta,
  gCvalues,
  spsMatch,
  ecozones,
  params3,
  params4,
  params5,
  params6,
  params7
)
}
\arguments{
\item{meta}{DESCRIPTION NEEDED}

\item{gCvalues}{DESCRIPTION NEEDED}

\item{spsMatch}{DESCRIPTION NEEDED}

\item{ecozones}{DESCRIPTION NEEDED}

\item{params3}{\code{data.frame} corresponding to Table 3 from Boudewyn et al. (2007),
available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table3.csv}.}

\item{params4}{\code{data.frame} corresponding to Table 4 from Boudewyn et al. (2007),
available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table4.csv}.}

\item{params5}{\code{data.frame} corresponding to Table 5 from Boudewyn et al. (2007),
available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table5.csv}.}

\item{params6}{\code{data.frame} corresponding to Table 6 from Boudewyn et al. (2007),
available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table6.csv}.}

\item{params7}{\code{data.frame} corresponding to Table 7 from Boudewyn et al. (2007),
available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table7.csv}.}
}
\value{
three-column matrix with columns corresponding to biomass (\eqn{T/ha}) for
total merchantable, foliage, and other.
}
\description{
Implements the flowchart from figure 3 of Boudewyn et al. (2007) to determined the
total above ground biomass (\eqn{T/ha}) from gross merchantable volume (\eqn{m^3/ha}).
}
\references{
Boudewyn, P., Song, X., Magnussen, S., & Gillis, M. D. (2007). Model-based, volume-to-biomass
conversion for forested and vegetated land in Canada (BC-X-411). Natural Resource Canada,
Pacific Forestry Centre. \url{https://cfs.nrcan.gc.ca/pubwarehouse/pdfs/27434.pdf}
}
