% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_cTransfersAlluvial.R
\name{cTransfersAlluvial}
\alias{cTransfersAlluvial}
\title{\code{cTransfersAlluvial}}
\usage{
cTransfersAlluvial(
  cTransfers,
  distMatrixID = NULL,
  distName = NULL,
  spuID = NULL,
  sw = NULL,
  nearMatches = TRUE,
  identical = !ask,
  ask = interactive(),
  ...
)
}
\arguments{
\item{cTransfers}{TODO}

\item{distMatrixID}{disturbance_matrix_id of the disturbance to plot.}

\item{distName}{disturbance name you wish to plot. Required only if \code{distMatrixID} is not provided.}

\item{spuID}{CBM-CFS3 spatial unit ID of the disturbance to plot. Required only if \code{distMatrixID} is not provided.}

\item{sw}{logical. Softwood (TRUE) or hardwood (FALSE). Required only if \code{distMatrixID} is not provided.}

\item{nearMatches}{logical. Allow for near matches; e.g. "clearcut" can match "clear-cut".}

\item{identical}{logical. Require identical matches.}

\item{ask}{logical.
If TRUE, prompt the user to choose the correct matches.
If FALSE, the function will look for a single match to each input.}

\item{...}{additional arguments to \code{\link{.matchSelect}}}
}
\value{
alluvialDist Alluvial plot of a disturbance in a specific spatial unit.
}
\description{
Maps proportions of carbon transfers across pools during a disturbance.
}
