% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_cTransfersAlluvial.R
\name{cTransfersAlluvial}
\alias{cTransfersAlluvial}
\title{\code{cTransfersAlluvial}}
\usage{
cTransfersAlluvial(cTransfers, distMatrixID = NA, distName = NA, spu = NA)
}
\arguments{
\item{cTransfers}{TODO}

\item{distMatrixID}{disturbance_matrix_id of the disturbance to plot. The user needs to specify this or \code{distName}.}

\item{distName}{disturbance name you wish to plot. The user needs to specify this or \code{distMatrixID}.}

\item{spu}{spatial unit ID of the disturbance to plot. Required only if \code{distName} is used.}
}
\value{
alluvialDist Alluvial plot of a disturbance in a specific spatial unit.
}
\description{
Maps proportions of carbon transfers across pools during a disturbance.
}
