% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_cumPoolsCreate.R
\name{cumPoolsCreate}
\alias{cumPoolsCreate}
\title{Create \code{cumPools} data.table}
\usage{
cumPoolsCreate(
  fullSpecies,
  gcMeta,
  userGcM3,
  stable3,
  stable4,
  stable5,
  stable6,
  stable7,
  thisAdmin
)
}
\arguments{
\item{fullSpecies}{DESCRIPTION NEEDED}

\item{gcMeta}{DESCRIPTION NEEDED}

\item{userGcM3}{DESCRIPTION NEEDED}

\item{stable3}{DESCRIPTION NEEDED}

\item{stable4}{DESCRIPTION NEEDED}

\item{stable5}{DESCRIPTION NEEDED}

\item{stable6}{DESCRIPTION NEEDED}

\item{stable7}{DESCRIPTION NEEDED}

\item{thisAdmin}{DESCRIPTION NEEDED}
}
\value{
\code{cumPools} data.table
}
\description{
Create \code{cumPools} data.table
}
