% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_cumPoolsCreate.R
\name{cumPoolsCreate}
\alias{cumPoolsCreate}
\title{Create \code{cumPools} data.table}
\usage{
cumPoolsCreate(
  fullSpecies,
  gcMeta,
  userGcM3,
  stable3,
  stable4,
  stable5,
  stable6,
  stable7,
  thisAdmin
)
}
\arguments{
\item{fullSpecies}{Study area species' names}

\item{gcMeta}{Growth curve metadata table}

\item{userGcM3}{Table of growth curve volume by age}

\item{stable3}{Boudewyn et al. 2007 stem wood biomass model parameters table for merchantable-sized trees}

\item{stable4}{Boudewyn et al. 2007 stem wood biomass model parameters table for non-merchantable-sized trees}

\item{stable5}{Boudewyn et al. 2007 stem wood biomass model parameters table for sapling-sized trees}

\item{stable6}{Boudewyn et al. 2007 stem wood biomass model parameters table for proportion model parameters}

\item{stable7}{Boudewyn et al. 2007 stem wood biomass model parameters table for caps on proportion models}

\item{thisAdmin}{Ecozone and spatial unit information table for the study area}
}
\value{
\code{cumPools} data.table
}
\description{
Create \code{cumPools} data.table
}
