% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_cumPoolsCreate.R
\name{cumPoolsCreate}
\alias{cumPoolsCreate}
\title{Create \code{cumPools} data.table}
\usage{
cumPoolsCreate(
  fullSpecies,
  gcMeta,
  userGcM3,
  stable3,
  stable4,
  stable5,
  stable6,
  stable7,
  thisAdmin
)
}
\arguments{
\item{fullSpecies}{Species names found in study area.}

\item{gcMeta}{Growth curve metadata table that links species with additional attributes.}

\item{userGcM3}{Growth curve metadata table that links species with additional attributes.}

\item{stable3}{Boudewyn table 3 subset to study area attributes.}

\item{stable4}{Boudewyn table 3 subset to study area attributes.}

\item{stable5}{Boudewyn table 3 subset to study area attributes.}

\item{stable6}{Boudewyn table 3 subset to study area attributes.}

\item{stable7}{Boudewyn table 3 subset to study area attributes.}

\item{thisAdmin}{Study area metadata}
}
\value{
\code{cumPools} data.table
}
\description{
Create \code{cumPools} data.table
}
