% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_cumPoolsCreateAGB.R
\name{cumPoolsCreateAGB}
\alias{cumPoolsCreateAGB}
\title{Convert total above ground biomass into 3 carbon pools (\eqn{T/ha}).
Implements the flowchart from figure 3 of Boudewyn et al. (2007) using an alternative
set of parameter to divide total above ground biomass (\eqn{T/ha}) into total merchantable
stemwood biomass (\eqn{T/ha}), foliage biomass (\eqn{T/ha}), and other wood biomass (\eqn{T/ha}).
Biomass to converted to carbon using the biomass to carbon conversion factor.}
\usage{
cumPoolsCreateAGB(AGB, bRateBiomassToCarbon = 0.5, ...)
}
\arguments{
\item{AGB}{\code{data.frame} with columns: \code{juris_id}, \code{ecozone}, \code{age}, \code{B}
and \code{species}, \code{speciesCode}, or \code{canfi_species}.}

\item{bRateBiomassToCarbon}{numeric. Conversion factor of biomass to carbon.}

\item{...}{arguments to \code{\link{convertAGB2pools}}}
}
\value{
\code{AGB} table updated by reference with additional columns
\code{merch}, \code{foliage}, and \code{other} with carbon (\eqn{T/ha}) in each above ground pool.
}
\description{
Implements the flowchart from figure 3 of Boudewyn et al. (2007) using an alternative
set of parameter to divide total above ground biomass (\eqn{T/ha}) into total merchantable
stemwood biomass (\eqn{T/ha}), foliage biomass (\eqn{T/ha}), and other wood biomass (\eqn{T/ha}).
}
\references{
Boudewyn, P., Song, X., Magnussen, S., & Gillis, M. D. (2007). Model-based, volume-to-biomass
conversion for forested and vegetated land in Canada (BC-X-411). Natural Resource Canada,
Pacific Forestry Centre. \url{https://cfs.nrcan.gc.ca/pubwarehouse/pdfs/27434.pdf}
}
