% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_cumPoolsCreateAGB.R
\name{cumPoolsCreateAGB}
\alias{cumPoolsCreateAGB}
\title{Convert total above ground biomass into 3 pools (\eqn{T/ha}).}
\usage{
cumPoolsCreateAGB(
  allInfoAGBin,
  pixGroupCol,
  bTable14tb,
  bTable6tb = "https://nfi.nfis.org/resources/biomass_models/appendix2_table6_tb.csv",
  bTable7tb = "https://nfi.nfis.org/resources/biomass_models/appendix2_table7_tb.csv",
  bRateBiomassToCarbon = 0.5
)
}
\arguments{
\item{allInfoAGBin}{\code{data.frame} with at least six following columns: \code{canfi_species},
\code{speciesCode}, \code{ecozone}, \code{juris_id}, \code{age}, \code{B} and a column for pixel group identifier.}

\item{pixGroupCol}{the name of the column in \code{allInfoAGBin} serving as the pixel group
identifier.}

\item{bTable14tb}{Total biomass to gross merchantable volume conversion.
An alternate table to use instead of Boudewyn et al. (2007) table 14
when total biomass per hectare in tonnes (tb = stem wood + stem bark + branches + foliage)
is the independent variable instead of gross merchantable volume per hectare.}

\item{bTable6tb}{Boudewyn et al. (2007) an alternative set of proportion model parameters
to use when total biomass per hectare in tonnes (tb = stem wood + stem bark + branches + foliage)
is the independent variable instead of gross merchantable volume per hectare.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table6_tb.csv}.}

\item{bTable7tb}{Boudewyn et al. (2007) an alternative set of caps on proportion models
to use in conjunction with the alternative set of proportion models
when total biomass per hectare in tonnes (tb = stem wood + stem bark + branches + foliage)
is the independent variable instead of gross merchantable volume per hectare.
Available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table7_tb.csv}.}

\item{bRateBiomassToCarbon}{numeric. Conversion factor of biomass to carbon.}
}
\value{
biomass (\eqn{T/ha}) in each above ground pool for each cohort per pixel group.
}
\description{
Implements the flowchart from figure 3 of Boudewyn et al. (2007) using an alternative
set of parameter to divide total above ground biomass (\eqn{T/ha}) into total merchantable
stemwood biomass (\eqn{T/ha}), foliage biomass (\eqn{T/ha}), and other wood biomass (\eqn{T/ha}).
}
\references{
Boudewyn, P., Song, X., Magnussen, S., & Gillis, M. D. (2007). Model-based, volume-to-biomass
conversion for forested and vegetated land in Canada (BC-X-411). Natural Resource Canada,
Pacific Forestry Centre. \url{https://cfs.nrcan.gc.ca/pubwarehouse/pdfs/27434.pdf}
}
