% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_cumPoolsCreateAGB.R
\name{cumPoolsCreateAGB}
\alias{cumPoolsCreateAGB}
\title{Convert total above ground biomass into 3 pools (\eqn{T/ha}).}
\usage{
cumPoolsCreateAGB(
  AGB,
  pixGroupCol,
  table6,
  table7,
  tableMerchantability,
  biomassToCarbonRate = 0.5
)
}
\arguments{
\item{AGB}{\code{data.frame} with the following columns:
\code{juris_id}, \code{ecozone}, \code{canfi_species}, \code{age}, \code{B} and a column for pixel group identifier.}

\item{pixGroupCol}{the name of the column in \code{AGB} serving as the pixel group identifier.}

\item{table6}{\code{data.frame} corresponding to Table 3 from Boudewyn et al. (2007),
available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table6_tb.csv}.}

\item{table7}{\code{data.frame} corresponding to Table 4 from Boudewyn et al. (2007),
available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table7_tb.csv}.}

\item{tableMerchantability}{\code{data.frame} Parameters to estimate the proportion
of stemwood that is merchantable, approximates the relationship between stemwood
biomass and nonmerchfactor predicted by equation 2 of Boudewyn et al., 2007.}

\item{biomassToCarbonRate}{Conversion factor of biomass to carbon}
}
\value{
\code{AGB} table updated by reference with additional columns
\code{merch}, \code{foliage}, and \code{other} with biomass (\eqn{T/ha}) in each above ground pool.
}
\description{
Implements the flowchart from figure 3 of Boudewyn et al. (2007) using an alternative
set of parameter to divide total above ground biomass (\eqn{T/ha}) into total merchantable
stemwood biomass (\eqn{T/ha}), foliage biomass (\eqn{T/ha}), and other wood biomass (\eqn{T/ha}).
}
\references{
Boudewyn, P., Song, X., Magnussen, S., & Gillis, M. D. (2007). Model-based, volume-to-biomass
conversion for forested and vegetated land in Canada (BC-X-411). Natural Resource Canada,
Pacific Forestry Centre. \url{https://cfs.nrcan.gc.ca/pubwarehouse/pdfs/27434.pdf}
}
