% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep_disturbanceRasters.R
\name{dataPrep_disturbanceRasters}
\alias{dataPrep_disturbanceRasters}
\title{CBM data prep: disturbanceRasters}
\usage{
dataPrep_disturbanceRasters(
  disturbanceRastersList,
  templateRast = NULL,
  year = NULL
)
}
\arguments{
\item{disturbanceRastersList}{list of \code{disturbanceRasters}.
Each \code{disturbanceRasters} item must be named by a 4 digit year
such that a single terra \code{\link[terra]{SpatRaster}} layer can be accessed
for each disturbance year (e.g.  \code{disturbanceRasters[["1"]][["2025"]]}).
This can either be a list named with disturbance event IDs such that every
non-NA raster value is considered an event of that type,
or a 1-length unnamed list of \code{disturbanceRasters} where the raster values
contain event IDs.}

\item{templateRast}{terra \code{\link[terra]{SpatRaster}}.
Template raster to align rasters with 'mode' resampling.}

\item{year}{digit or character.
One or more years to summarize disturbance events for.
If NULL, all available years are summarized.}
}
\value{
\code{disturbanceEvents}
data.table with integer columns 'pixelIndex', 'year', 'eventID'.
}
\description{
Process a list of \code{disturbanceRasters} into \code{disturbanceEvents}.
}
