% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc-extractToRast.R
\name{extractToRast}
\alias{extractToRast}
\title{Extract to raster}
\usage{
extractToRast(input, templateRast, index = 1, crop = TRUE)
}
\arguments{
\item{input}{terra SpatRaster, one or more raster files, or sf polygons.}

\item{templateRast}{SpatRaster. Raster template.}

\item{index}{numeric or character. Raster layer or vector field to extract.}

\item{crop}{logical. Crop input before alignment to template raster.}
}
\value{
vector with a value for each cell of \code{templateRast}.
Data type matches input data type.
}
\description{
Extract values from a spatial data source for each cell of a template raster.
Resampling is done using "mode" resampling to preserve categorical values.
}
