% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumPoolsCreateAGB.R
\name{getParameters}
\alias{getParameters}
\title{Extract the parameters to apply to convert total biomass into pool biomass}
\usage{
getParameters(table6, table7, canfi_species, ecozone, juris_id)
}
\arguments{
\item{table6}{\code{data.frame} corresponding to Table 3 from Boudewyn et al. (2007),
available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table6_tb.csv}.}

\item{table7}{\code{data.frame} corresponding to Table 4 from Boudewyn et al. (2007),
available from \url{https://nfi.nfis.org/resources/biomass_models/appendix2_table7_tb.csv}.}

\item{canfi_species}{the canfi code of the species}

\item{ecozone}{the code of the ecozone}

\item{juris_id}{the 2-letter code for the province/territory}
}
\value{
a list with 2 vectors for the parameters in table6 and table7 respectively.
}
\description{
Extract the species- and location- specific parameters for equation 4-7 of
Boudewyn et al. (2007). If there is no match for the given ecozone, the parameters
for a different ecozone in the same province/territory is returned. If there
is no match for a given province/territory, the parameters for a different
province/territory in the same ecozone is returned. If there is no match for
the given ecozone and province/territory, the parameters for a different location
is returned.
}
\references{
Boudewyn, P., Song, X., Magnussen, S., & Gillis, M. D. (2007). Model-based, volume-to-biomass
conversion for forested and vegetated land in Canada (BC-X-411). Natural Resource Canada,
Pacific Forestry Centre. \url{https://cfs.nrcan.gc.ca/pubwarehouse/pdfs/27434.pdf}
}
