% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-CFS3_DB-ListDisturbances.R
\name{histDist}
\alias{histDist}
\title{Historical disturbances}
\usage{
histDist(mySpu = c(27, 28))
}
\arguments{
\item{mySpu}{Numeric spatial unit id(s).}
}
\description{
Identifies the stand-replacing wildfire disturbance in each spatial unit.
}
\details{
Historical disturbances in CBM-CFS3 are used for "filling-up" the soil-related carbon pools.
Boudewyn et al. (2007) translate the m3/ha curves into biomass per ha in each of four pools:
total biomass for stem wood, total biomass for bark, total biomass for branches and total
biomass for foliage.
Biomass in coarse and fine roots, in aboveground- and belowground- very-fast, -fast, -slow,
in medium-soil, and in snags still needs to be estimated.
In all spatial units in Canada, the historical disturbance is set to fire.
A stand-replacing fire disturbance is used in a disturb-grow cycle, where stands are disturbed
and regrown with turnover, overmature, decay, functioning until the dead organic matter pools
biomass values stabilize (+/- 10\%).
\subsection{TODO: (I think but that is in the Rcpp-RCMBGrowthIncrements.cpp so can't check).}{

By default the most recent is selected, but the user can change that.
}
}
