% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_m3ToBiomPlots.R
\name{m3ToBiomPlots}
\alias{m3ToBiomPlots}
\title{Plot all columns that are not id_col}
\usage{
m3ToBiomPlots(inc, id_col = "gcids")
}
\arguments{
\item{inc}{DESCRIPTION NEEDED}

\item{id_col}{DESCRIPTION NEEDED}
}
\description{
Plot all columns that are not id_col
}
