% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_mapNPP.R
\name{mapNPP}
\alias{mapNPP}
\title{\code{mapNPP}}
\usage{
mapNPP(flux, masterRaster, year = NULL)
}
\arguments{
\item{flux}{data.table. Table of flux for each cohort or for each pixel.
Must include the 'pixelIndex' column.}

\item{masterRaster}{\code{\link[terra]{SpatRaster}} template defining the study area.}

\item{year}{numeric. Year that the map represents.
If provided, it will be included in the plot title.}
}
\value{
\code{ggplot}
}
\description{
Map net primary productivity (NPP) across a study area.
}
