% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matchSelect.R
\name{.matchSelect}
\alias{.matchSelect}
\title{Match select}
\usage{
.matchSelect(
  inputs,
  choices,
  allowNA = FALSE,
  identical = TRUE,
  nearMatches = list(),
  funSimplify = function(x) trimws(tolower(unname(as.character(x)))),
  ask = !identical & interactive(),
  inputTable = NULL,
  choiceTable = NULL,
  choiceTableExtra = NULL
)
}
\arguments{
\item{inputs}{character. Strings to match}

\item{choices}{character. Match choices}

\item{allowNA}{logical. Return NA when no match is found.
Otherwise an error will be thrown if any input cannot be matched.}

\item{identical}{logical. Require identical matches.}

\item{nearMatches}{list. A list of near matches to allow.
Item names must match \code{inputs} and item contents must be vectors of additional allowable matches.}

\item{funSimplify}{function.
Method for simplifying input and choices strings before matching.
By default, the strings will be cast to lower case and white space is trimmed.}

\item{ask}{logical.
If TRUE, prompt the user to choose the correct matches.
If FALSE, the function will look for a single match to each input.}

\item{inputTable}{data.frame. If \code{ask}, include a table of input attributes
to print to the user during the selection prompt.}

\item{choiceTable}{data.frame. If \code{ask}, include a table of input attributes
to print to the user during the selection prompt.}
}
\value{
integer. Match indexes
}
\description{
Match select
}
\keyword{internal}
