% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{NPPplot}
\alias{NPPplot}
\title{NPPplot}
\usage{
NPPplot(spatialDT, NPP, masterRaster)
}
\arguments{
\item{spatialDT}{TODO}

\item{masterRaster}{TODO}

\item{changeInNPP}{TODO}

\item{time}{TODO}
}
\value{
TODO
}
\description{
NPPplot
}
