% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{NPPPlot}
\alias{NPPPlot}
\title{NPPPlot}
\usage{
NPPPlot(spatialDT, changeInNPP, masterRaster, time)
}
\arguments{
\item{spatialDT}{TODO}

\item{changeInNPP}{TODO}

\item{masterRaster}{TODO}

\item{time}{TODO}
}
\value{
TODO
}
\description{
NPPPlot
}
