% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots.R
\name{NPPplot}
\alias{NPPplot}
\title{\code{NPPplot}}
\usage{
NPPplot(cohortGroupKeep, NPP, masterRaster)
}
\arguments{
\item{cohortGroupKeep}{TODO}

\item{NPP}{TODO}

\item{masterRaster}{A \code{SpatRaster} object to use as the template}
}
\value{
TODO
}
\description{
\code{NPPplot}
}
