% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_plotEmissionsProducts.R
\name{plotEmissionsProducts}
\alias{plotEmissionsProducts}
\title{\code{plotEmissionsProducts}}
\usage{
plotEmissionsProducts(emissionsProducts)
}
\arguments{
\item{emissionsProducts}{Table of simlation emissions and products by cohort group by year}
}
\value{
\code{ggplot}
}
\description{
Plot yearly emissions and products.
}
