% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputs.R
\name{prepInputsEcozones}
\alias{prepInputsEcozones}
\title{\code{prepInputsEcozones}}
\usage{
prepInputsEcozones(url, dPath, rasterToMatch)
}
\arguments{
\item{url}{A URL to the data}

\item{dPath}{destination path}

\item{rasterToMatch}{A raster with NAs representing "off study area" and otherwise
any arbitrary value for "in study area".
Equivalent to \code{rasterToMatch} argument in \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}.}
}
\description{
\code{prepInputsEcozones}
}
