% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boudewyn_cumPoolsCreateAGB.R
\name{propMerch}
\alias{propMerch}
\title{Calculates the proportion of total stemwood considered merchantable.}
\usage{
propMerch(totalStemWood, age, params)
}
\arguments{
\item{totalStemWood}{\code{numeric} Total Stemwood vector.}

\item{age}{\code{numeric} Age of the cohort.}

\item{params}{\code{data.table} Parameters to calculate propotion of merchantable stemwood.}
}
\value{
\code{numeric} Vector of the proportion of stemwood that is merchantable.
}
\description{
Calculates the proportion of total stemwood considered merchantable.
}
