% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Python_ReticulateFindPython.R
\name{reticulate_install_python_windows}
\alias{reticulate_install_python_windows}
\title{Install Python with reticulate::install_python}
\usage{
reticulate_install_python_windows(
  version = NULL,
  useGit = TRUE,
  prompt = interactive(),
  pyenvRoot = tools::R_user_dir("CBMutils")
)
}
\arguments{
\item{version}{character. Python version string.}

\item{useGit}{logical. Allow reticulate to clone pyenv-win if Git is available}

\item{prompt}{logical. Prompt user to approve download of pyenv-win tool}

\item{pyenvRoot}{character. Path to directory of where to download the pyenv-win tool}
}
\description{
Download the pyenv-win Python version management tool from Github if necessary.
See: https://github.com/pyenv-win/pyenv-win
}
