% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadesCBMextraFunctions.R
\name{retrieveSpuRaster}
\alias{retrieveSpuRaster}
\title{Produce a raster with `spUnits`}
\usage{
retrieveSpuRaster(
  spatialUnitsFile = shapefile("data/spUnit_Locator.shp"),
  UserArea,
  rasterRes = c(250, 250)
)
}
\arguments{
\item{spatialUnitsFile}{DESCRIPTION NEEDED}

\item{UserArea}{DESCRIPTION NEEDED}

\item{rasterRes}{The desired raster resolution.}
}
\description{
Produce a raster with `spUnits`
}
\examples{
\dontrun{
  test1 <- raster::shapefile("data/forIan/SK_data/CBM_GIS/SpadesCBM_TestArea.shp")
  out1 <- retrieveSpuRaster(UserArea = test1, rasterRes = c(250, 250))
  if (interactive()) Plot(out1)

  if (require(LandR)) {
    test2 <- LandR::randomStudyArea(seed = 100, size = 10000 * 100 * 300)
    out2 <- retrieveSpuRaster(UserArea = test2, rasterRes = c(250, 250))
    if (interactive())Plot(out2)
  }
}

}
