% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthCurves.R
\name{sapfac}
\alias{sapfac}
\title{sapfac}
\usage{
sapfac(table5, eq2)
}
\arguments{
\item{table5}{DESCRIPTION NEEDED}

\item{eq2}{DESCRIPTION NEEDED}
}
\description{
eq3 is for the saplings and it needs b_nm from the previous eq2
}
