% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-CFS3_DB-ListDisturbances.R
\name{seeDist}
\alias{seeDist}
\title{See disturbances}
\usage{
seeDist(
  EXN = TRUE,
  matrixIDs = NULL,
  dbPath = NULL,
  disturbance_matrix_value = NULL
)
}
\arguments{
\item{EXN}{logical. Use CBM-EXN CBM-CFS3 equivalent model data.}

\item{matrixIDs}{character. Optional. Subset disturbances by disturbance_matrix_id}

\item{dbPath}{Path to CBM-CFS3 SQLite database file.
Required if EXN = FALSE}

\item{disturbance_matrix_value}{disturbance_matrix_value table from CBM-EXN
Required if EXN = TRUE}
}
\value{
List of \code{data.frame} named by disturbance_matrix_id
}
\description{
Retrieve disturbance source pools, sink pools, and the proportions transferred.
}
