% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-CFS3_DB-ListDisturbances.R
\name{seeDist}
\alias{seeDist}
\title{See disturbances}
\usage{
seeDist(distId, dbPath)
}
\arguments{
\item{distId}{Description needed}

\item{dbPath}{Path to sqlite database file.}
}
\value{
A list of \code{data.frame}s, one per disturbance matrix id.
}
\description{
Get the descriptive name of the disturbance, the source pools, the sink pools, and
the proportions transferred.
}
