% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-CBM_core_dbReadSummary.R
\name{simCBMdbReadSummary}
\alias{simCBMdbReadSummary}
\title{simList: Read SpaDES CBM database summary}
\usage{
simCBMdbReadSummary(
  simCBM,
  summary,
  by = "cohortID",
  year = NULL,
  years = NULL,
  useCache = TRUE
)
}
\arguments{
\item{simCBM}{SpaDES.core \code{\link[SpaDES.core]{simList}} for a CBM simulation including the \strong{CBM_core} module.}

\item{summary}{character. Name of summary table to collect.}

\item{by}{character. Name of field to summarize results by.
Data can be summarized by 'cohortID', 'pixelIndex', or 'year'}

\item{year}{numeric. Year of simulation results.
Required if \verb{by = 'cohortID' or 'pixelIndex'}.
Defaults to the current simulation year.}

\item{years}{numeric. Years of simulation results to summarize.
Required if \code{by = 'year'}.
Defaults to the full simulation time span.}

\item{useCache}{logical. Cache database query result to the database's
\code{cache} directory with \code{\link[reproducible]{Cache}}.}
}
\value{
data.table with key matching the \code{by} argument.
}
\description{
Read a data summary from a simulation's \strong{CBM_core} SpaDES CBM database.
}
