% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-CBM_core_dbReadTable.R
\name{simCBMdbReadTable}
\alias{simCBMdbReadTable}
\title{simList: Read SpaDES CBM database table}
\usage{
simCBMdbReadTable(simCBM, year, table, useCache = TRUE)
}
\arguments{
\item{simCBM}{SpaDES.core \code{\link[SpaDES.core]{simList}} for a CBM simulation including the \strong{CBM_core} module.}

\item{year}{numeric. Year of simulation results.
Defaults to the current simulation year.}

\item{table}{character. Name of database table to read.
One of 'key', 'parameters', 'state', 'flux', or 'pools'.}

\item{useCache}{logical. Cache database query result to the database's
\code{cache} directory with \code{\link[reproducible]{Cache}}.}
}
\value{
data.table with key 'cohortID',
columns 'pixelIndex' and 'row_idx' (cohort group ID),
and any additional columns in the source table.
}
\description{
Read a table of cohort data from a simulation's \strong{CBM_core} SpaDES CBM database.
}
