% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-CBM_core_dbWrite.R
\name{simCBMdbWrite}
\alias{simCBMdbWrite}
\title{simList: Write SpaDES CBM database}
\usage{
simCBMdbWrite(
  simCBM,
  year,
  parameters = TRUE,
  state = TRUE,
  flux = TRUE,
  pools = TRUE
)
}
\arguments{
\item{simCBM}{SpaDES.core \code{\link[SpaDES.core]{simList}} for a CBM simulation including the \strong{CBM_core} module.}

\item{year}{numeric. Simulation data year.
Defaults to the current simulation year.}

\item{parameters}{logical. Write the 'parameters' table to file.}

\item{state}{logical. Write the 'state' table to file.}

\item{flux}{logical. Write the 'flux' table to file.}

\item{pools}{logical. Write the 'pools' table to file.}
}
\description{
Write simulation \strong{CBM_core} module data to the SpaDES CBM database.
}
