% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_mapNPP.R
\name{simMapNPP}
\alias{simMapNPP}
\title{\code{simMapNPP}}
\usage{
simMapNPP(simCBM, year, cbm4_results = NULL, useCache = TRUE)
}
\arguments{
\item{simCBM}{SpaDES.core \code{\link[SpaDES.core]{simList}} for a CBM simulation including the \strong{CBM_core} module.}

\item{year}{numeric. Year of simulation results.}

\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{CBM4::cbm4_results_processor}.}

\item{useCache}{logical. Cache database query result to the database's
\code{cache} directory with \code{\link[reproducible]{Cache}}.}
}
\value{
\code{ggplot}
}
\description{
Map net primary productivity (NPP) across a study area.
}
