% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_plotEmissionsProducts.R
\name{simPlotEmissionsProducts}
\alias{simPlotEmissionsProducts}
\title{\code{simPlotEmissionsProducts}}
\usage{
simPlotEmissionsProducts(
  simCBM,
  years = NULL,
  cbm4_results = NULL,
  useCache = TRUE
)
}
\arguments{
\item{simCBM}{SpaDES.core \code{\link[SpaDES.core]{simList}} for a CBM simulation including the \strong{CBM_core} module.}

\item{years}{numeric. Simulation years to include in plot. Defaults to all simulation years.}

\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{CBM4::cbm4_results_processor}.}

\item{useCache}{logical. Cache database query result to the database's
\code{cache} directory with \code{\link[reproducible]{Cache}}.}
}
\value{
\code{ggplot}
}
\description{
Plot yearly emissions and products.
}
