% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots_plotEmissionsProducts.R
\name{simPlotEmissionsProducts}
\alias{simPlotEmissionsProducts}
\title{\code{simPlotEmissionsProducts}}
\usage{
simPlotEmissionsProducts(simCBM)
}
\arguments{
\item{simCBM}{SpaDES.core \code{\link[SpaDES.core]{simList}} for a CBM simulation including the \strong{CBM_core} module.}
}
\value{
\code{ggplot}
}
\description{
Plot yearly emissions and products.
}
