% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-CBM_core_dbReadTable.R
\name{spadesCBMdbReadTable}
\alias{spadesCBMdbReadTable}
\title{Read SpaDES CBM database table}
\usage{
spadesCBMdbReadTable(spadesCBMdb, year, table, useCache = TRUE)
}
\arguments{
\item{spadesCBMdb}{\strong{CBM_core} module outputs database path.}

\item{year}{numeric. Year of simulation results.}

\item{table}{character. Name of database table to read.
One of 'key', 'parameters', 'state', 'flux', or 'pools'.}

\item{useCache}{logical. Cache database query result to the database's
\code{cache} directory with \code{\link[reproducible]{Cache}}.}
}
\value{
data.table with key 'cohortID',
columns 'pixelIndex' and 'row_idx' (cohort group ID),
and any additional columns in the source table.
}
\description{
Read a table of cohort data from a \strong{CBM_core} SpaDES CBM database.
}
