% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-CBM_core_dbWrite.R
\name{spadesCBMdbWrite}
\alias{spadesCBMdbWrite}
\title{Write SpaDES CBM database}
\usage{
spadesCBMdbWrite(
  spadesCBMdb,
  cbm_vars,
  year,
  parameters = TRUE,
  state = TRUE,
  flux = TRUE,
  pools = TRUE
)
}
\arguments{
\item{spadesCBMdb}{\strong{CBM_core} module outputs database path.}

\item{parameters}{logical. Write the 'parameters' table to file.}

\item{state}{logical. Write the 'state' table to file.}

\item{flux}{logical. Write the 'flux' table to file.}

\item{pools}{logical. Write the 'pools' table to file.}
}
\description{
Write \strong{CBM_core} module data to the SpaDES CBM database.
}
\keyword{internal}
