% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{spatialPlot}
\alias{spatialPlot}
\title{\code{spatialPlot}}
\usage{
spatialPlot(pixelkeep, cbmPools, poolsToPlot, years, masterRaster)
}
\arguments{
\item{pixelkeep}{TODO}

\item{cbmPools}{TODO}

\item{poolsToPlot}{TODO}

\item{years}{TODO}

\item{masterRaster}{TODO}
}
\value{
TODO
}
\description{
\code{spatialPlot}
}
