% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-plots.R
\name{spatialPlot}
\alias{spatialPlot}
\title{\code{spatialPlot}}
\usage{
spatialPlot(cbmPools, year, masterRaster, cohortGroupKeep)
}
\arguments{
\item{cbmPools}{TODO}

\item{year}{TODO}

\item{masterRaster}{A \code{SpatRaster} object to use as the template}

\item{cohortGroupKeep}{TODO}
}
\value{
TODO
}
\description{
Maps the total carbon across the landscape for a given year.
}
