% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-DB_species.R
\name{sppMatch}
\alias{sppMatch}
\title{Species match}
\usage{
sppMatch(
  species,
  matchCol = NULL,
  sppEquivalencies = NULL,
  checkNA = c("CBM_speciesID", "Broadleaf")
)
}
\arguments{
\item{species}{Species identifiers.}

\item{matchCol}{character. \code{sppEquivalencies} columns to match \code{species} with.
Defaults to \code{LandR::sppEquivalencies_CA} columns with Latin and generic English species names.}

\item{sppEquivalencies}{data.table. Table of species identifiers and metadata.
Defaults to \code{LandR::sppEquivalencies_CA}.}

\item{checkNA}{character. \code{sppEquivalencies} columns to check for NA values in.}
}
\value{
data.table. Subset of \code{sppEquivalencies} with 1 row per species.
}
\description{
Retrieve species metadata by matching species names or other identifiers with columns in \code{sppEquivalencies}.
}
