% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-CFS3_DB-ListDisturbances.R
\name{spuDist}
\alias{spuDist}
\title{CBM-CFS3 Spatial Unit Disturbances}
\usage{
spuDist(spuIDs, dbPath, localeID = 1)
}
\arguments{
\item{spuIDs}{Spatial unit ID(s)}

\item{dbPath}{Path to CBM-CFS3 SQLite database file}

\item{localeID}{CBM-CFS3 locale_id}
}
\value{
\code{data.table} with columns
'spatial_unit_id', 'disturbance_type_id', 'disturbance_matrix_id',
'name', 'description'
}
\description{
Identify the disturbances possible in spatial units.
}
