% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadesCBMextraFunctions.R
\name{spuDist}
\alias{spuDist}
\title{Identify the ID number (CBM-CFS3 legacy) possible in the current spatial unit}
\usage{
spuDist(mySpu, dbPath)
}
\arguments{
\item{mySpu}{Numeric spatial unit id(s).}

\item{dbPath}{Path to sqlite database file.}
}
\description{
You give is spatial units you are targeting \code{mySpu} and it gives you the disturbance matrix id
that are possible/default in that specific spu and a descriptive name of that disturbance matrix
it creates a \code{data.frame} of length number of disturbances, with three columns:
\code{spatial_unit_is}, \code{disturbance_matrix_id}, and a \code{desciption} of the disturbance.
}
\details{
TODO: can we have a Canada-wide SPU map and they locate themselves on the map?
this needs to be done before simulations are run so the user can provide this
info (location info) for the simulations - Ian is working on this.

the function has the defaults from the SK managed forest example.
These can be changed by feeding in other SPU.
}
\examples{
\dontrun{
  ## using raster
  library(raster)
  spuRaster <- raster(file.path("data/forIan/SK_data/CBM_GIS/spUnits_TestArea.tif"))
  spatial_unit_id <- getValues(spuRaster) # 28 27
  mySpu <- unique(spatial_unit_id)

  ## using growth curves
  f <- file.path("spadesCBMinputs/data/SK_ReclineRuns30m/LookupTables/yieldRCBM.csv")
  gcIn <- as.matrix(read.csv(f))
  mySpu <- unique(gcIn[, 1])
}
}
