% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBM-CFS3_DB-ListDisturbances.R
\name{spuDistMatch}
\alias{spuDistMatch}
\title{CBM-CFS3 Disturbances Match}
\usage{
spuDistMatch(
  distTable,
  ask = interactive(),
  nearMatches = TRUE,
  listDist = NULL,
  ...
)
}
\arguments{
\item{distTable}{\code{data.table} with columns 'spatial_unit_id' and 'name' (or 'distName').
The name column will be matched with disturbance names and descriptions
in the CBM-CFS3 database.}

\item{ask}{logical.
If TRUE, prompt the user to choose the correct disturbance matches.
If FALSE, the function will look for exact name matches.}

\item{nearMatches}{logical. Allow for near matches; e.g. "clearcut" can match "clear-cut".}

\item{listDist}{data.table. Optional. Result of a call to \code{\link{spuDist}}.
A list of possible disturbances in the spatial unit(s) with columns
'spatial_unit_id', 'disturbance_type_id', 'disturbance_matrix_id', 'name', 'description'.}

\item{...}{arguments to \code{\link{spuDist}}
for listing the possible disturbances in the spatial units.}
}
\value{
\code{data.table} with columns 'spatial_unit_id'
'disturbance_type_id', 'disturbance_matrix_id', 'name', 'description'
}
\description{
Match disturbance names with CBM-CFS3 spatial unit disturbances.
}
